% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/actor.count.R
\name{actor.count}
\alias{actor.count}
\title{Number of actors in an affiliation network}
\usage{
actor.count(bigraph)
}
\arguments{
\item{bigraph}{The ambient affiliation network}
}
\description{
This function returns the number of actors (nodes of type FALSE) in an
affiliation network.
}
\examples{
data(inner.circle)
actor.count(inner.circle)
}

