% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/actor.projection.R
\name{actor.projection}
\alias{actor.projection}
\title{Projection onto actors of an affiliation network}
\usage{
actor.projection(bigraph, name = "name")
}
\arguments{
\item{bigraph}{The ambient affiliation network}

\item{name}{What attribute of the actor nodes in bigraph to name the nodes
in the projection (defaults to 'name')}
}
\description{
This function uses the igraph function bipartite.projection to compute the
projection of an affiliation network onto the actor nodes.
}
\examples{
data(hobson.inner.circle)
tab <- table(V(hobson.inner.circle)$type)
proj <- actor.projection(hobson.inner.circle)
vcount(proj) == tab[1]
}

