% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actor-count.R
\name{actor_count}
\alias{actor_count}
\title{Count the actors in an affiliation network}
\usage{
actor_count(bigraph)
}
\arguments{
\item{bigraph}{An affiliation network; see \code{\link{is_an}}.}
}
\description{
This function returns the number of actors (nodes of type FALSE) in an
affiliation network.
}
\examples{
data(southafrica1905)
actor_count(southafrica1905)
}

