% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actor-projection.R
\name{actor_projection}
\alias{actor_projection}
\title{Project an affiliation network onto its actors}
\usage{
actor_projection(bigraph, name = "name")
}
\arguments{
\item{bigraph}{An affiliation network; see \code{\link{is_an}}.}

\item{name}{What attribute of the actor nodes in bigraph to name the nodes in
the projection (defaults to "name")}
}
\description{
This function uses the igraph function \code{\link{bipartite_projection}} to
compute the projection of an affiliation network onto the actor nodes.
}
\examples{
data(southafrica1905)
tab <- table(V(southafrica1905)$type)
proj <- actor_projection(southafrica1905)
vcount(proj) == tab[1]
}

