% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/an.dual.R
\name{an.dual}
\alias{an.dual}
\title{Dual of an affiliation network}
\usage{
an.dual(bigraph)
}
\arguments{
\item{bigraph}{The ambient affiliation network}
}
\description{
This function swaps the TRUE and FALSE values of the `type` attribute,
which in `bitriad` has the effect of taking the dual of the affiliation
network (swapping the roles of actors and events).
}
\examples{
data(hobson.inner.circle)
tab <- table(V(hobson.inner.circle)$type)
proj <- actor.projection(an.dual(hobson.inner.circle))
vcount(proj) == tab[2]
}

