% Generated by roxygen2 (4.0.1): do not edit by hand
\name{an.triad}
\alias{an.triad}
\title{Affiliation network triads}
\usage{
an.triad(lambda, w, actors = letters[16:18], events = if (sum(c(lambda, w))
  == 0) c() else as.character(1:sum(c(lambda, w))))
}
\arguments{
\item{lambda}{The vector of pairwise weights}

\item{w}{The triadwise weight}

\item{actors}{Actor names (defaults to 'P', 'Q', and 'R')}

\item{events}{Event names (defaults to positive integers)}
}
\description{
Affiliation network triads
}

