% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/anPlotSpecs.R
\name{anPlotSpecs}
\alias{anPlotSpecs}
\title{Parameter specs for plotting affiliation networks}
\usage{
anPlotSpecs(bigraph)
}
\arguments{
\item{bigraph}{An affiliation network.}
}
\description{
This function assigns several parameters to a bipartite graph to prepare it
for visualization using `plot.igraph`.
}
\examples{
data(davis.clique)
data(whigs)
for(g in list(davis.clique, whigs)) {
    plot(anPlotSpecs(g))
}
}

