% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/census.projection.R
\name{census.projection}
\alias{census.projection}
\title{Project a triad census}
\usage{
census.projection(census, scheme = NULL)
}
\arguments{
\item{census}{The input triad census. It is classified as "uniformity",
"structural", or "simple" if its dimensons are 8-by-2, 4-by-2, or 4-by-1,
respectively, or unless otherwise specified; otherwise it is classified as
"full".}

\item{scheme}{The type of census the inputed census should be taken to be.}
}
\description{
This function inputes a triad census of any scheme and returns a list of
triad censuses projected from it (not icluding itself).
}

