% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transitivity.census.R
\name{census.transitivity}
\alias{census.transitivity}
\title{Global transitivity from a triad census}
\usage{
census.transitivity(census, flavor, scheme = NULL, openFun, closedFun,
  stat = "clust", counts = FALSE)
}
\arguments{
\item{census}{Numeric matrix; the input triad census. It is classified as
"uniformity", "structural", or "simple" if its dimensons are 8-by-2, 4-by-2,
or 4-by-1, respectively, or unless otherwise specified; otherwise it is
classified as "full".}

\item{flavor}{Character; the type of transitivity (matched to
"watts.strogatz", "classical", "opsahl", "exclusive", "allact", "indequ",
"indstr", "injact", "injequ", or "injstr")}

\item{scheme}{Character; the type of census (to be matched to "full",
"uniformity", "structural", or "simple").}

\item{openFun}{The open wedge count for a triad (ignored if `flavor` is not
`NULL`).}

\item{closedFun}{The closed wedge count for a triad (ignored if `flavor` is
not `NULL`).}

\item{stat}{Character; the form of the statistic (matched to "clustering" or
"transitivity"; defaults to "clust"; ignored if `counts` is `TRUE`).}

\item{counts}{Logical; whether to return open and closed wedge counts instead
of a ratio statistic (if `TRUE`, overrides `stat`; defaults to `FALSE`).}
}
\description{
Each global clustering coefficient can be recovered from the full triad
census, and some from smaller censuses. This function determines the type of
census being inputed and computes the desired measure of global transitivity,
if possible, otherwise throwing an error.
}

