% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/centeredTriads.R
\name{centeredTriads}
\alias{centeredTriads}
\title{Classes for all triads centered at a given set of nodes}
\usage{
centeredTriads(bigraph, vids)
}
\arguments{
\item{bigraph}{An affiliation network; see \code{is.an}.}

\item{vids}{A subset of actor node ids.}
}
\value{
A 5-column matrix of centered triads, each described by its center and exclusive and inclusive event counts.
}
\description{
Classes for all triads centered at a given set of nodes
}

