% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{chicago1960s}
\alias{chicago1960s}
\title{Corporate interlocks among directors in Chicago.}
\format{
An affiliation network; see \link{is_an}.
}
\source{
Barnes, R. & Burkett, T. (2010). Structural Redundancy and
Multiplicity in Corporate Networks. \emph{Connections}, 30(2), p. 4--20.
\url{https://www.insna.org/connections-archives-2000-2015}
}
\description{
An affiliation network of 20 directors of 24 corporations and social clubs.
}
