% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/combinatorial-bijections.R
\name{index_subset}
\alias{index_subset}
\alias{subset_index}
\alias{subset_partition}
\alias{partition_subset}
\alias{index_partition}
\alias{partition_index}
\alias{combinatorial_bijections}
\alias{index_subset_R}
\alias{subset_index_R}
\alias{subset_partition_R}
\alias{partition_subset_R}
\alias{index_partition_R}
\alias{partition_index_R}
\alias{indexSubset}
\alias{indexPartition}
\alias{subsetIndex}
\alias{subsetPartition}
\alias{partitionIndex}
\alias{partitionSubset}
\title{Combinatorial bijections for affiliation network triad indexing}
\usage{
index_subset(i)

subset_index(vec)

subset_partition(vec)

partition_subset(lambda)

index_partition(i)

partition_index(lambda)

index_subset_R(i)

subset_index_R(vec)

subset_partition_R(vec)

partition_subset_R(lambda)

index_partition_R(i)

partition_index_R(lambda)

indexSubset(i)

indexPartition(i)

subsetIndex(vec)

subsetPartition(vec)

partitionIndex(lambda)

partitionSubset(lambda)
}
\arguments{
\item{i}{An index in the total order; starts at 0.}

\item{vec}{A (decreasing) 3-subset of natural numbers (including 0).}

\item{lambda}{A partition of at most 3 parts.}
}
\description{
These functions biject among partitions of at most 3 parts, 3-subsets of
natural numbers, and indices for the lexicographic total orders on them.
}
\examples{
index_subset_R(2)
index_partition_R(2)
subset_index_R(c(3, 2, 0))
subset_partition_R(c(3, 2, 0))
partition_index_R(c(1, 1, 0))
partition_subset_R(c(1, 1, 0))
}
