% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coop.triad.census}
\alias{coop.triad.census}
\title{Cooperativity triad census}
\usage{
coop.triad.census(bigraph, type = 0, verbose = FALSE)
}
\arguments{
\item{bigraph}{The affiliation network}

\item{type}{The actor node type in bigraph}

\item{verbose}{Logical; whether to display progress bars}
}
\description{
This function computes the cooperativity triad census for an affiliation
network, a 4-by-2 array of frequency counts for each isomorphism class of
triad, modulo an equivalence relation of event nodes by which actors they
are tied to, which sum to choose(n, 3), where n is the number of actor nodes.
}

