% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ctc2stc}
\alias{ctc2stc}
\title{Cooperativity triad census to simple triad census}
\usage{
ctc2stc(ctc)
}
\arguments{
\item{ctc}{A cooperativity two-mode triad census (not necessarily labeled)}
}
\description{
Cooperativity triad census to simple triad census
}

