% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{davis.group}
\alias{davis.group}
\title{Group of women connected by event coattendance in Old City.}
\format{An undirected igraph object with nodes of logical attribute `type`;
women correspond to nodes of type `FALSE`, events to nodes of type `TRUE`.
Events also carry a `time` attribute equal to the number of days since the
previous 31 December.}
\source{
Davis, Gardner, and Gardner (1941) *Deep South: A Social
Anthropological Study of Caste and Class*
\url{http://books.google.com/books?id=HGIdAAAAIAAJ}
}
\description{
A dynamic affiliation network of 18 women who attended 14 events over a 9-
month period.
}

