% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{ddggs.clique}
\alias{ddggs.clique}
\title{Group of women connected by event coattendance in Old City.}
\format{An undirected igraph object with nodes of logical attribute `type`;
women correspond to nodes of type `FALSE`, activities to nodes of type
`TRUE`.}
\source{
Davis, Gardner, and Gardner (1941) *Deep South: A Social
Anthropological Study of Caste and Class*
\url{http://books.google.com/books?id=HGIdAAAAIAAJ}
}
\description{
A dynamic affiliation network of 5 women who participated in 5 activities.
}

