% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dual.an.R
\name{dual.an}
\alias{dual.an}
\title{Dual of an affiliation network}
\usage{
dual.an(bigraph)
}
\arguments{
\item{bigraph}{The ambient affiliation network}
}
\description{
This function swaps the TRUE and FALSE values of the `type` attribute,
which in `bitriad` has the effect of taking the dual of the affiliation
network (swapping the roles of actors and events).
}
\examples{
data(inner.circle)
tab <- table(V(inner.circle)$type)
proj <- actor.projection(dual.an(inner.circle))
vcount(proj) == tab[2]
}

