% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual-an.R
\name{dual_an}
\alias{dual_an}
\title{Take the dual of an affiliation network}
\usage{
dual_an(bigraph)
}
\arguments{
\item{bigraph}{An affiliation network; see \code{\link{is_an}}.}
}
\description{
This function negates the logical values of the vertex \code{type} attribute
and reorders the vertex ids accordingly, effectively obtaining the dual 
affiliation network.
}
\examples{
data(southafrica1905)
tab <- table(igraph::V(southafrica1905)$type)
proj <- actor_projection(dual_an(southafrica1905))
igraph::vcount(proj) == tab[2]
}

