% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dyn.transitivity.an.R
\name{dyn.transitivity.an}
\alias{dyn.transitivity.an}
\title{Dynamic triadic closure of an affiliation network}
\usage{
dyn.transitivity.an(graph, memory = Inf, type = "global")
}
\arguments{
\item{memory}{Numeric; a duration of time after which events are forgotten.}

\item{type}{Character; whether to compute the global or local statistic, or
to return a 2-column matrix of wedge counts (defaults to "global").}

\item{bigraph}{An affiliation network with time-stamped events.}
}
\description{
Given an affiliation network with time-stamped events, compute the proportion
of centered triples at which an open wedge exists at some time that is
closed at a later time.
}
\examples{
data(women.group)
dyn.transitivity.an(women.group)
cbind(
    transitivity(actor.projection(women.group), type = "local"),
    opsahl.transitivity(women.group, type = "local"),
    excl.transitivity(women.group, type = "local"),
    dyn.transitivity.an(women.group, type = "local")
)
}

