% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dyn.transitivity.bigraph.R
\name{dyn.transitivity.bigraph}
\alias{dyn.transitivity.bigraph}
\title{Dynamic triadic closure}
\usage{
dyn.transitivity.bigraph(graph, memory = Inf, type = "global",
  count.closures = FALSE)
}
\arguments{
\item{memory}{Numeric; a duration of time after which events are forgotten}

\item{type}{Character; whether to compute the global or local statistic, or
to return a 2-column matrix of wedge counts (defaults to `global`)}

\item{count.closures}{Logical; whether to count each closure of a wedge once
(FALSE) or as many times as the wedge is closed (TRUE)}

\item{bigraph}{An affiliation network with time-stamped events}
}
\description{
Given an affiliation network with time-stamped events, compute the proportion
of centered triples at which an open wedge exists at some time that is
closed at a later time.
}

