% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dyn.transitivity.graph.R
\name{dyn.transitivity.graph}
\alias{dyn.transitivity.graph}
\title{Dynamic triadic closure}
\usage{
dyn.transitivity.graph(graph, memory)
}
\arguments{
\item{graph}{A simple network with time-stamped edges}

\item{memory}{Numeric; a duration of time after which edges are forgotten}
}
\description{
Given a simple network with time-stamped edges, compute the proportion
of centered triples at which an open wedge exists at some time that is
closed at a later time.
}

