% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dyn.transitivity.R
\name{dyn.transitivity}
\alias{dyn.transitivity}
\title{Dynamic triadic closure}
\usage{
dyn.transitivity(graph, memory = Inf, type = "global",
  count.closures = FALSE)
}
\arguments{
\item{graph}{An affiliation network with time-stamped events or a simple
network with time-stamped edges}

\item{memory}{Numeric; a duration of time after which events or edges are
forgotten. Defaults to Inf}
}
\description{
Given an affiliation network with time-stamped events, or a simple network
with time-stamped edges, compute the proportion of centered triples at which
an open wedge exists at some time that is closed at a later time.
}
\examples{
data(davis.group)
dyn.transitivity(davis.group)
cbind(
    transitivity(actor.projection(davis.group), type = 'local'),
    opsahl.transitivity(davis.group, type = 'local'),
    excl.transitivity(davis.group, type = 'local'),
    dyn.transitivity(davis.group, type = 'local')
)
}

