% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dyn.triadic.closure.bigraph}
\alias{dyn.triadic.closure.bigraph}
\title{Dynamic triadic closure}
\usage{
dyn.triadic.closure.bigraph(graph, memory = Inf, type = "global",
  count.closures = FALSE)
}
\arguments{
\item{memory}{Numeric; a duration of time after which events are forgotten}

\item{bigraph}{An affiliation network with time-stamped events}
}
\description{
Given an affiliation network with time-stamped events, compute the proportion
of centered triples at which an open wedge exists at some time that is
closed at a later time.
}

