% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-transitivity-an.R
\name{dynamic-transitivity}
\alias{dynamic-transitivity}
\alias{dynamic_transitivity_an}
\alias{dyn.transitivity.an}
\title{Dynamic triadic closure of an affiliation network}
\usage{
dynamic_transitivity_an(bigraph, memory = Inf, type = "global")

dyn.transitivity.an(bigraph, memory = Inf, type = "global")
}
\arguments{
\item{bigraph}{An affiliation network with time-stamped events.}

\item{memory}{Numeric; a duration of time after which events are forgotten.}

\item{type}{Character; whether to compute the global or local statistic, or
to return a 2-column matrix of wedge counts (defaults to "global").}
}
\description{
Given an affiliation network with time-stamped events, compute the proportion
of centered triples at which an open wedge exists at some time that is
closed at a later time.
}
\examples{
data(women_group)
dynamic_transitivity_an(women_group)
cbind(
    transitivity(actor_projection(women_group), type = "local"),
    triad_closure_opsahl(women_group, type = "local"),
    triad_closure_exclusive(women_group, type = "local"),
    dynamic_transitivity_an(women_group, type = "local")
)
}
