% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-triad-closure-an.R
\name{dynamic_triad_closure}
\alias{dynamic_triad_closure}
\alias{dynamic_triad_closure_an}
\alias{dynamic_transitivity_an}
\alias{dyn.transitivity.an}
\alias{dynamic_triad_closure_projection}
\title{Triadic closure for dynamic affiliation networks}
\usage{
dynamic_triad_closure_an(bigraph, actors = V(bigraph)[V(bigraph)$type ==
  FALSE], type = "global", ..., measure = NULL)

dynamic_transitivity_an(bigraph, actors = V(bigraph)[V(bigraph)$type ==
  FALSE], type = "global", ..., measure = NULL)

dyn.transitivity.an(bigraph, actors = V(bigraph)[V(bigraph)$type == FALSE],
  type = "global", ..., measure = NULL)

dynamic_triad_closure_projection(bigraph, memory = Inf, type = "global")
}
\arguments{
\item{bigraph}{An affiliation network with time-stamped events.}

\item{actors}{A vector of actor nodes in \code{bigraph}.}

\item{type}{The type of statistic, matched to \code{"global"}, 
\code{"local"}, or \code{"raw"}.}

\item{...}{Additional parameters passed to specific functions.}

\item{measure}{Character; the measure of triad closure, used as the suffix 
\code{*} to \code{triad_closure_*}. Matched to \code{"classical"} (also 
\code{"watts_strogatz"}), \code{"twomode"} (also \code{"opsahl"}), 
\code{"unconnected"} (also \code{"liebig_rao_0"}), 
\code{"completely_connected"} (also \code{"liebig_rao_3"}), 
\code{"exclusive"}, or \code{"projection"}.}

\item{memory}{Numeric; a duration of time after which events are forgotten.}
}
\description{
Given an affiliation network with time-stamped events, compute the proportion
of centered triples at which an open wedge exists at some time that is closed
at a later time.
}
\examples{
data(women_group)
dynamic_triad_closure_an(women_group)
cbind(
  transitivity(actor_projection(women_group), type = "local"),
  triad_closure_opsahl(women_group, type = "local"),
  triad_closure_exclusive(women_group, type = "local"),
  dynamic_triad_closure_an(women_group, type = "local")
)
}
