% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/event.projection.R
\name{event.projection}
\alias{event.projection}
\title{Project an affiliation network onto its events}
\usage{
event.projection(bigraph, name = "name")
}
\arguments{
\item{bigraph}{An affiliation network; see \code{\link{is.an}}.}

\item{name}{What attribute of the event nodes in bigraph to name the nodes in
the projection (defaults to "name")}
}
\description{
This function uses the igraph function \code{\link{bipartite.projection}} to
compute the projection of an affiliation network onto the event nodes.
}
\examples{
data(southafrica1905)
tab <- table(V(southafrica1905)$type)
proj <- event.projection(southafrica1905)
vcount(proj) == tab[2]
}

