% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-count.R
\name{event_count}
\alias{event_count}
\title{Count the events in an affiliation network}
\usage{
event_count(bigraph)
}
\arguments{
\item{bigraph}{An affiliation network; see \code{\link{is_an}}.}
}
\description{
This function returns the number of events (nodes of type TRUE) in an
affiliation network.
}
\examples{
data(southafrica1905)
event_count(southafrica1905)
}

