% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{fischer.han.whigs}
\alias{fischer.han.whigs}
\title{Membership network of American Whigs.}
\format{An undirected igraph object with nodes of logical attribute `type`;
members correspond to nodes of type `FALSE`, organizations to nodes of type
`TRUE`.}
\source{
Fischer (1994) *Paul Revere's Ride*
\url{https://github.com/kjhealy/revere}
}
\description{
An affiliation network of 136 members of 5 American Whig organizations.
}

