% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ftc2CXw.R
\name{ftc2CXw}
\alias{ftc2CXw}
\title{Global clustering coefficients from the triad census}
\usage{
ftc2CXw(ftc, ww, num.denom = FALSE, by.tri = FALSE)
}
\arguments{
\item{ftc}{A full triad census (a matrix)}

\item{num.denom}{Whether to return the numerator and denominator of the
clustering coefficient as a length-2 vector.}

\item{S.fn}{The "success" function, the closed wedge count for a triad}

\item{F.fn}{The "failure" function, the unclosed wedge count for a triad}
}
\description{
Each global clustering coefficient can be recovered from the full triad
census. the first function provides a framework for this calculation, and the
following call that framework for specific clustering coefficients.
}

