% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ftc2indequ.R
\name{ftc2indequ}
\alias{ftc2indequ}
\title{Global clustering coefficients from the triad census}
\usage{
ftc2indequ(census)
}
\arguments{
\item{census}{Numeric matrix; a full triad census}
}
\description{
Global clustering coefficients from the triad census
}

