% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ftc2indstr.R
\name{ftc2indstr}
\alias{ftc2indstr}
\title{Global clustering coefficients from the triad census}
\usage{
ftc2indstr(census)
}
\arguments{
\item{census}{Numeric matrix; a full triad census}
}
\description{
Global clustering coefficients from the triad census
}

