% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ftc2injequ}
\alias{ftc2injequ}
\title{Global clustering coefficients from the triad census}
\usage{
ftc2injequ(ftc, num.denom = FALSE, by.tri = FALSE)
}
\arguments{
\item{ftc}{A full triad census (a matrix)}

\item{S.fn}{The "success" function, the closed wedge count for a triad}

\item{F.fn}{The "failure" function, the unclosed wedge count for a triad}

\item{num.denom}{Whether to return the numerator and denominator of the
clustering coefficient as a length-2 vector.}
}
\description{
Each global clustering coefficient can be recovered from the full triad
census. the first function provides a framework for this calculation, and the
following call that framework for specific clustering coefficients.
}

