% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ftc2stc.R
\name{ftc2stc}
\alias{ftc2stc}
\title{Full triad census to structural triad census}
\usage{
ftc2stc(tc)
}
\arguments{
\item{tc}{A full triad census (not necessarily labeled)}
}
\description{
TO BE DELETED
}

