% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ftc2utc}
\alias{ftc2utc}
\title{Full triad census to uniformity triad census}
\usage{
ftc2utc(tc)
}
\arguments{
\item{tc}{A full two-mode triad census (not necessarily labeled)}
}
\description{
Full triad census to uniformity triad census
}

