% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{galaskiewicz.ceos}
\alias{galaskiewicz.ceos}
\title{Interlocks among corporate philanthropists.}
\format{An undirected igraph object with nodes of logical attribute `type`;
directors correspond to nodes of type `FALSE`, companies to nodes of type
`TRUE`.}
\source{
Wasserman & Faust (1994) *Social Network Analysis*
\url{http://books.google.com/books?id=CAm2DpIqRUIC}
}
\description{
An affiliation network of 26 directors of 15 companies, taken as a subset
from a larger dataset.
}

