% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hallCriterion.R
\name{hallCriterion}
\alias{hallCriterion}
\title{Hall criterion}
\usage{
hallCriterion(lst)
}
\arguments{
\item{lst}{A list of vectors possibly having some common elements}
}
\description{
A criterion for the existence of a system of distinct representatives for a
finite set:
www.encyclopediaofmath.org/index.php/System_of_different_representatives
}

