% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{hobson.inner.circle}
\alias{hobson.inner.circle}
\title{Small inner ring of South African finance.}
\format{An undirected igraph object with nodes of logical attribute `type`;
directors correspond to nodes of type `FALSE`, companies to nodes of type
`TRUE`.}
\source{
Hobson, John A. (1919) *The Evolution of Modern Capitalism*
\url{https://archive.org/details/evolutionofmoder00hobsuoft}
}
\description{
An affiliation network of six multiple directors of five gold, diamond, and
finance companies.
}

