% Generated by roxygen2 (4.0.1): do not edit by hand
\name{incr.triad.census}
\alias{incr.triad.census}
\title{Incremental triad census}
\usage{
incr.triad.census(bigraph, type = 0)
}
\arguments{
\item{bigraph}{The affiliation network}

\item{type}{The actor node type in bigraph}
}
\description{
This function computes the incremental triad census for an affiliation
network, an 8-by-2 array of frequency counts for each isomorphism class of
triad, modulo the equivalence relation that identifies triads based on the
positivity of their triad weights w and of the differences between their
sequential pairwise weights. This census can be recovered from the full triad
census, and the structural equivalence triad census can be recovered from it.
The census must sum to choose(n, 3), where n is
the number of actor nodes.
}

