% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/indexSubset.R
\name{indexSubset}
\alias{indexSubset}
\title{Combinatorial bijections for affiliation network triad labeling}
\usage{
indexSubset(i)
}
\arguments{
\item{i}{An index under the total order. Starts at 0.}
}
\description{
These functions biject among partitions of at most 3 parts, 3-subsets of
natural numbers, and indices for the lexicographic total orders on both.
}
\examples{
indexSubset(2)
}

