% Generated by roxygen2 (4.0.1): do not edit by hand
\name{injact.wedges}
\alias{injact.wedges}
\title{Wedges}
\usage{
injact.wedges(bigraph, Q)
}
\arguments{
\item{bigraph}{The ambient affiliation network.}

\item{Q}{An actor node in the network.}
}
\description{
Each clustering coefficient can be defined as the proportion of "wedges" that
are "closed", for suitable definitions of both terms. These functions count
the "wedges", and among them the "closed" ones, centered at a given actor
node in a given affiliation network.
}

