% Generated by roxygen2 (4.0.1): do not edit by hand
\name{injequ.transitivity}
\alias{injequ.transitivity}
\title{Affiliation network clustering coefficients}
\usage{
injequ.transitivity(bigraph, node.type = 0, type = "global",
  vids = which(V(bigraph)$type == node.type))
}
\arguments{
\item{bigraph}{An affiliation network.}

\item{node.type}{The actor node type in the bigraph object (defaults to 0)}

\item{type}{The type of clustering coefficient (defaults to 'global')}

\item{wedges.fn}{The wedge function (see the entry on "Wedges")}

\item{vids}{A subset of actor node ids at which to evaluate the local
clustering coefficient.}
}
\description{
Each clustering coefficient can be defined as the proportion of "wedges" that
are "closed", for suitable definitions of both terms. The main function,
an.transitivity, calls one of the wedge functions and computes the
global or local clustering coefficient of the given affiliation network,
and if the local, then at the given nodes.
}

