% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.an.R
\name{is.an}
\alias{is.an}
\title{Test and coerce affiliation network conditions}
\usage{
is.an(graph)
}
\arguments{
\item{graph}{An igraph object}
}
\description{
These functions test for or, when possible, impose the properties used in
bitriad to distinguish affiliation networks from the very general igraph
class.
}

