% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.dyn.R
\name{is.dyn}
\alias{is.dyn}
\title{Dynamic graphs}
\usage{
is.dyn(graph)
}
\arguments{
\item{graph}{An igraph object.}
}
\description{
A simple graph is dynamic if its edges have time stamps. A bipartite graph is
dynamic if its event nodes have time stamps.
}

