% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is.dynamic.R
\name{is.dynamic}
\alias{is.dynamic}
\title{Dynamic graphs}
\usage{
is.dynamic(graph)
}
\arguments{
\item{graph}{An igraph object}
}
\description{
A simple graph is dynamic if its edges have time stamps. A bipartite graph is
dynamic if its nodes (most importantly, its event nodes) have time stamps.
}

