% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is.triad.R
\name{is.triad}
\alias{is.triad}
\title{Affiliation network triads}
\usage{
is.triad(graph)
}
\arguments{
\item{graph}{An affiliation network; see `is.an`.}
}
\description{
Affiliation network triads
}

