% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-dynamic.R
\name{is_dynamic}
\alias{is_dynamic}
\title{Dynamic graphs}
\usage{
is_dynamic(graph)
}
\arguments{
\item{graph}{An igraph object.}
}
\description{
A simple graph is dynamic if its edges have time stamps. A bipartite graph is
dynamic if its event nodes have time stamps.
}

