% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/layout.triad.R
\name{layout.triad}
\alias{layout.triad}
\title{Affiliation network triads}
\usage{
layout.triad(lambda, w, scale = 0.3, angdir = -1, rot = -pi/2,
  rot.lambda = c(0, 0, 0), rot.w = pi/12)
}
\arguments{
\item{lambda}{The vector of pairwise weights}

\item{w}{The triadwise weight}
}
\description{
Affiliation network triads
}

