% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{minneapolis1970s}
\alias{minneapolis1970s}
\title{Interlocks among corporate philanthropists in Minneapolis-St. Paul.}
\format{An affiliation network; see \code{\link{is_an}}.}
\source{
Wasserman & Faust (1994) *Social Network Analysis*
\url{http://books.google.com/books?id=CAm2DpIqRUIC}
}
\description{
These data record the memberships of 26 directors on the boards of 15
companies. They are a subset of a larger dataset that is not publicly
available.
}
