% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode-counts.R
\name{mode-counts}
\alias{mode-counts}
\alias{actor_count}
\alias{event_count}
\alias{actor.count}
\alias{event.count}
\title{Count the actors and events in an affiliation network}
\usage{
actor_count(bigraph)

event_count(bigraph)

actor.count(bigraph)

event.count(bigraph)
}
\arguments{
\item{bigraph}{An affiliation network.}
}
\description{
These functions return the number of actors (nodes with
  \code{type} attribute \code{FALSE}) or events (\code{TRUE}) in an
  affiliation network.
}
\examples{
data(chicago1960s)
actor_count(chicago1960s)
event_count(chicago1960s)
}
