% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode-projection.r
\name{mode-projection}
\alias{mode-projection}
\alias{mode_projection}
\alias{actor_projection}
\alias{event_projection}
\alias{actor.projection}
\alias{event.projection}
\title{Project an affiliation network onto its actors}
\usage{
mode_projection(bigraph, mode = "actors", name = "name")

actor_projection(bigraph, name = "name")

event_projection(bigraph, name = "name")

actor.projection(bigraph, name = "name")

event.projection(bigraph, name = "name")
}
\arguments{
\item{bigraph}{An affiliation network.}

\item{mode}{Numeric or character; whether to project onto actors (\code{1} or
\code{"actors"}) or onto events (\code{2} or \code{"events"}).}

\item{name}{Character; the attribute of the actor or event nodes in 
\code{bigraph} to use as names for the nodes in the projection.}
}
\description{
These functions use \code{\link{bipartite_projection}} to 
  compute the projections of an affiliation network onto the actor or event 
  nodes.
}
\examples{
data(chicago1960s)
tab <- table(V(chicago1960s)$type)
proj <- actor_projection(chicago1960s)
vcount(proj) == tab[1]
proj <- event_projection(chicago1960s)
vcount(proj) == tab[2]
}
