% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode-counts.R
\name{mode_counts}
\alias{mode_counts}
\alias{actor_count}
\alias{event_count}
\alias{actor.count}
\alias{event.count}
\title{Count the actors and events in an affiliation network}
\usage{
actor_count(bigraph)

event_count(bigraph)

actor.count(bigraph)

event.count(bigraph)
}
\arguments{
\item{bigraph}{An affiliation network; see \code{\link{is_an}}.}
}
\description{
These functions return the number of actors (nodes of type "FALSE") or events
("TRUE") in an affiliation network.
}
\examples{
data(southafrica1905)
actor_count(southafrica1905)
event_count(southafrica1905)
}
