% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode-projection.r
\name{mode_projection}
\alias{mode_projection}
\alias{actor_projection}
\alias{event_projection}
\alias{actor.projection}
\alias{event.projection}
\title{Project an affiliation network onto its actors}
\usage{
mode_projection(graph, mode = 1, name = "name")

actor_projection(graph, name = "name")

event_projection(graph, name = "name")

actor.projection(graph, name = "name")

event.projection(graph, name = "name")
}
\arguments{
\item{graph}{An affiliation network.}

\item{mode}{Numeric or character; whether to project onto actors (\code{1} or
\code{"actors"}) or onto events (\code{2} or \code{"events"}).}

\item{name}{Character; the attribute of the actor or event nodes in 
\code{graph} to use as names for the nodes in the projection. If \code{NA},
node IDs are converted to characters and used. If \code{NULL}, no names are
assigned.}
}
\description{
These functions use \code{\link{bipartite_projection}} to 
  compute the projections of an affiliation network onto the actor or event 
  nodes.
}
\examples{
data(chicago1960s)
tab <- table(V(chicago1960s)$type)
proj <- actor_projection(chicago1960s)
vcount(proj) == tab[1]
proj <- event_projection(chicago1960s)
vcount(proj) == tab[2]
}
\seealso{
Original \strong{igraph} functions: 
  \code{\link[igraph]{bipartite_projection}}
}
