% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode-projections.R
\name{mode_projections}
\alias{mode_projections}
\alias{actor_projection}
\alias{event_projection}
\alias{actor.projection}
\alias{event.projection}
\title{Project an affiliation network onto its actors}
\usage{
actor_projection(bigraph, name = "name")

event_projection(bigraph, name = "name")

actor.projection(bigraph, name = "name")

event.projection(bigraph, name = "name")
}
\arguments{
\item{bigraph}{An affiliation network; see \code{\link{is_an}}.}

\item{name}{What attribute of the actor nodes in bigraph to name the nodes in
the projection (defaults to "name")}
}
\description{
These functions use the igraph function \code{\link{bipartite_projection}} to
compute the projections of an affiliation network onto the actor and event
nodes.
}
\examples{
data(chicago1960s)
tab <- table(V(chicago1960s)$type)
proj <- actor_projection(chicago1960s)
vcount(proj) == tab[1]
proj <- event_projection(chicago1960s)
vcount(proj) == tab[2]
}
