% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{nmt.meetings}
\alias{nmt.meetings}
\title{Noordin Top meeting attendance network.}
\format{An affiliation network; see \code{\link{is.an}}.}
\source{
The Association of Religious Data Archives
\url{http://www.thearda.com/Archive/Files/Descriptions/TERRNET.asp}
}
\description{
These data record the attendance of 26 individuals at 20 meetings
associated with Noordin Mohammad Top.
}

