% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/opsahl.transitivity.R
\name{opsahl.transitivity}
\alias{opsahl.transitivity}
\title{Affiliation network clustering coefficients}
\usage{
opsahl.transitivity(bigraph, type = "global",
  vids = which(!V(bigraph)$type))
}
\arguments{
\item{bigraph}{An affiliation network.}

\item{type}{The type of clustering coefficient (defaults to 'global')}

\item{vids}{A subset of actor node ids at which to evaluate the local
clustering coefficient.}
}
\description{
Each clustering coefficient can be defined as the proportion of "wedges" that
are "closed", for suitable definitions of both terms. The main function,
transitivity.an, calls one of the wedge functions and computes the
global or local clustering coefficient of the given affiliation network,
and if the local, then at the given nodes.
}
\examples{
data(women.group)
opsahl.transitivity(women.group)
cbind(
    transitivity(actor.projection(women.group), type = 'local'),
    opsahl.transitivity(women.group, type = 'local'),
    excl.transitivity(women.group, type = 'local')
)
}

