% Generated by roxygen2 (4.0.1): do not edit by hand
\name{partition.subset}
\alias{partition.subset}
\title{Combinatorial bijections for affiliation triad labeling}
\usage{
partition.subset(lambda)
}
\arguments{
\item{lambda}{A partition of at most 3 parts.}
}
\description{
These functions biject among partitions of at most 3 parts, 3-subsets of
natural numbers, and indices for the lexicographic total orders on both.
}
\examples{
partition.subset(c(1, 1, 0))
}

