% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/partitionIndex.R
\name{partitionIndex}
\alias{partitionIndex}
\title{Combinatorial bijections for affiliation network triad indexing}
\usage{
partitionIndex(lambda)
}
\arguments{
\item{lambda}{A partition of 3 parts (possibly including 0).}
}
\value{
An index in the total order; starts at 0.
}
\description{
These functions biject among partitions of at most 3 parts, 3-subsets of
natural numbers, and indices for the lexicographic total orders on them.
}
\examples{
partitionIndex(c(1, 1, 0))
}

