% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/partitionSubset.R
\name{partitionSubset}
\alias{partitionSubset}
\title{Combinatorial bijections for affiliation network triad labeling}
\usage{
partitionSubset(lambda)
}
\arguments{
\item{lambda}{A partition of at most 3 parts.}
}
\description{
These functions biject among partitions of at most 3 parts, 3-subsets of
natural numbers, and indices for the lexicographic total orders on both.
}
\examples{
partitionSubset(c(1, 1, 0))
}

