% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.an.R
\name{plot.an}
\alias{plot.an}
\title{Plot affiliation networks}
\usage{
\method{plot}{an}(bigraph, layoutFun = layout.bipartite, ...)
}
\arguments{
\item{bigraph}{An affiliation network.}

\item{layoutFun}{Layout function to use (defaults to `layout.bipartite`).}

\item{...}{Additional parameters passed to layoutFun}
}
\description{
This function passes affiliation network specifications to `plot.igraph`.
}

