% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prettify.an.R
\name{prettify.an}
\alias{prettify.an}
\title{Set plotting parameters for an affiliation network}
\usage{
prettify.an(bigraph)
}
\arguments{
\item{bigraph}{An affiliation network; see `is.an`.}
}
\description{
This function assigns several plotting parameters to the vertices and edges
of a bipartite graph to prepare it for visualization using `plot.igraph`.
}
\examples{
data(women.clique)
data(whigs)
for(g in list(women.clique, whigs)) {
    plot(prettify.an(g))
}
}

