% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify-an.R
\name{prettify}
\alias{prettify}
\alias{prettify_an}
\alias{prettify.an}
\title{Set plotting parameters for an affiliation network}
\usage{
prettify_an(bigraph)

prettify.an(bigraph)
}
\arguments{
\item{bigraph}{An affiliation network; see \code{is_an}.}
}
\description{
This function assigns several plotting parameters to the vertices and edges 
of a bipartite graph to prepare it for visualization using
\code{plot.igraph}.
}
\examples{
library(igraph)
data(women_clique)
data(whigs)
for(g in list(women_clique, whigs)) {
    plot(prettify_an(g))
}
}
