% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-census.R
\name{project_census}
\alias{ftc2stc}
\alias{ftc2tc}
\alias{ftc2utc}
\alias{project.census}
\alias{project_census}
\alias{stc2tc}
\alias{utc2stc}
\alias{utc2tc}
\title{Project a triad census}
\usage{
project_census(census, scheme = NULL, add.names = FALSE)

project.census(census, scheme = NULL, add.names = FALSE)

ftc2utc(census)

ftc2stc(census)

ftc2tc(census)

utc2stc(census)

utc2tc(census)

stc2tc(census)
}
\arguments{
\item{census}{Numeric matrix or vector; the input triad census. It is
classified as "uniformity", "structural", or "simple" if its dimensons are
8-by-2, 4-by-2, or 4-by-1, respectively, or unless otherwise specified;
otherwise it is classified as "full".}

\item{scheme}{Character; the type of census (to be matched to "full",
"uniformity", "structural", or "simple").}

\item{add.names}{Logical; whether to label the rows and (where applicable)
columns of the projected censuses.}
}
\description{
This function inputes an affiliation network triad census of any scheme and
returns a list of triad censuses projected from it (not icluding itself). In
order of projectability, the census schemes are full (affiliation network),
uniformity, structural, and simple.
}

