% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-transitivity.R
\name{project_transitivity}
\alias{project_transitivity}
\alias{project.transitivity}
\title{Affiliation network clustering coefficients}
\usage{
project_transitivity(graph, type = "global", vids = which(!V(graph)$type))

project.transitivity(graph, type = "global", vids = which(!V(graph)$type))
}
\arguments{
\item{graph}{An affiliation network.}

\item{type}{The type of clustering coefficient (defaults to "global")}

\item{vids}{A subset of actor node ids at which to evaluate the local
clustering coefficient.}
}
\description{
Each clustering coefficient can be defined as the proportion of "wedges" that
are "closed", for suitable definitions of both terms. The main function,
transitivity_an, calls one of the wedge functions and computes the
global or local clustering coefficient of the given affiliation network,
and if the local, then at the given nodes. (`project_transitivity`
cheats by using the native `transitivity` but produces output consistent
with the other variants of `transitivity_an`.)
}
\seealso{
Other triad closure functions: \code{\link{dynamic_triad_closure}},
  \code{\link{transitivity_an}},
  \code{\link{triad_closure_from_census}},
  \code{\link{triad_closure}}
}
