% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{scotland1920s}
\alias{scotland1920s}
\title{Networks of interlocking directorates}
\format{An affiliation network; see \code{\link{is_an}}.}
\source{
Scott, J., & Hughes, M. (1980). \emph{The Anatomy of Scottish
Capital}. Croom Helm. 
\url{http://books.google.com/books?id=59mvAwAAQBAJ}
}
\description{
Five affiliation networks constructed from the multiple directors lists of
the top >100 Scottish companies over five two-year intervals between 1904 and
1974.
}
