% Generated by roxygen2 (4.0.1): do not edit by hand
\name{share.weight}
\alias{share.weight}
\title{Schedule}
\usage{
share.weight(bigraph, vids, name = "name")
}
\arguments{
\item{bigraph}{The ambient affiliation network}

\item{graph}{A one-mode network (used as the one-mode projection of bigraph)}

\item{vids}{Vertex ids, the event count among which is being considered}

\item{vp}{Ids of vertex pair, the edge between which is being considered}

\item{v}{A subset of actor node ids.}
}
\description{
The schedule of a subset of actor nodes in an affiliation network is the
induced subgraph on the actor nodes together with all event nodes incident
to at least two of the actor nodes.
}

