% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simple.triad.census}
\alias{simple.triad.census}
\title{Simple triad census}
\usage{
simple.triad.census(graph)
}
\arguments{
\item{graph}{A one-mode network}
}
\description{
This function inputs a simple graph and returns a vector of frequency counts
for the four isomorphism classes of triad, which therefore sum to
choose(n, 3), where n is the number of nodes. This function is needed because
igraph::triad.census, from which in small cases a simple triad census can be
recovered through strategic coordinate sums, returns nonsense frequencies for
many large graphs.
}

