% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-triad-census.R
\name{simple_triad_census}
\alias{simple_triad_census}
\alias{simple.triad.census}
\title{Simple triad census}
\usage{
simple_triad_census(graph, add.names = FALSE)

simple.triad.census(graph, add.names = FALSE)
}
\arguments{
\item{graph}{A simple, undirected \code{igraph} object.}

\item{add.names}{Logical; whether to label the matrix rows and columns.}
}
\description{
Given a simple graph, return a vector indicating the number of 
  triads of each isomorphism class, namely whether the triad contains zero, 
  one, two, or three links. The values must sum to \eqn{n!/(3!(n-3)!)},
  where \eqn{n} is the number of nodes.
}
