% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{southafrica1905}
\alias{southafrica1905}
\title{Small inner ring of South African finance.}
\format{An affiliation network; see `is.an`.}
\source{
Hobson, John A. (1919) *The Evolution of Modern Capitalism*
\url{https://archive.org/details/evolutionofmoder00hobsuoft}
}
\description{
These data record the memberships of 6 directors on the boards of 5 gold,
diamond, and finance companies.
}

