% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stc2tc.R
\name{stc2tc}
\alias{stc2tc}
\title{Structural triad census to simple triad census}
\usage{
stc2tc(tc)
}
\arguments{
\item{tc}{A structural triad census (not necessarily labeled); 4 by 2 matrix}
}
\description{
Structural triad census to simple triad census
}

