% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str.triad.census.R
\name{str.triad.census}
\alias{str.triad.census}
\title{Structural triad census}
\usage{
\method{str}{triad.census}(bigraph, verbose = FALSE)
}
\arguments{
\item{bigraph}{The affiliation network}

\item{verbose}{Logical; whether to display progress bars}
}
\description{
This function computes the structural triad census for an affiliation
network, a 4-by-2 array of frequency counts for each isomorphism class of
triad, modulo an equivalence relation of event nodes by which actors they
are tied to, which sum to choose(n, 3), where n is the number of actor nodes.
}

