% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/subsetIndex.R
\name{subsetIndex}
\alias{subsetIndex}
\title{Combinatorial bijections for affiliation network triad labeling}
\usage{
subsetIndex(vec)
}
\arguments{
\item{vec}{A (decreasing) 3-subset of natural numbers (including 0).}
}
\description{
These functions biject among partitions of at most 3 parts, 3-subsets of
natural numbers, and indices for the lexicographic total orders on both.
}
\examples{
subsetIndex(c(3, 2, 0))
}

