% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sync.mat.R
\name{sync.mat}
\alias{sync.mat}
\title{Matrix dimension synchronization}
\usage{
sync.mat(lst)
}
\arguments{
\item{lst}{A list of matrices}
}
\description{
Given a list of matrices, extend them all by the necessary zero rows to have
dimension the greatest number of rows by the greatest number of columns among
them.
}

