% Generated by roxygen2 (4.0.2): do not edit by hand
\name{three.tied.triads}
\alias{three.tied.triads}
\title{Triad tallies}
\usage{
three.tied.triads(bigraph, type = 0, graph = actor.projection(bigraph, type
  = type, name = "id"))
}
\arguments{
\item{bigraph}{The ambient affiliation network from which graph is projected}

\item{type}{The actor node type in bigraph}

\item{graph}{A one-mode network}
}
\description{
These functions are called by the affiliation network triad census to handle
triads of different types using the projection onto actors. The name of each
function indicates the number of edges that appear among the three actors of
the triad in the projection. (Zero-edge triads do not need to be tallied;
the total number of triads is easily calculated, and the difference between
this number and the total number of triads with edges gives the number of
triads without.)
}

