% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tmtc2ctc}
\alias{tmtc2ctc}
\title{Two-mode triad census to cooperativity triad census}
\usage{
tmtc2ctc(tmtc)
}
\arguments{
\item{tmtc}{A full two-mode triad census (not necessarily labeled)}
}
\description{
Two-mode triad census to cooperativity triad census
}

