% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tmtc2CXw}
\alias{tmtc2CXw}
\title{Global clustering coefficients from the two-mode triad census}
\usage{
tmtc2CXw(tc, ww, num.denom = FALSE, by.tri = FALSE)
}
\arguments{
\item{tc}{A two-mode triad census (a matrix)}

\item{S.fn}{The "success" function, the closed wedge count for a triad}

\item{F.fn}{The "failure" function, the unclosed wedge count for a triad}

\item{num.denom}{Whether to return the numerator and denominator of the
clustering coefficient as a length-2 vector.}
}
\description{
Each global clustering coefficient can be recovered from the two-mode triad
census. the first function provides a framework for this calculation, and the
following call that framework for specific clustering coefficients.
}

