% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tmtc2itc}
\alias{tmtc2itc}
\title{Two-mode triad census to incremental triad census}
\usage{
tmtc2itc(tmtc)
}
\arguments{
\item{tmtc}{A full two-mode triad census (not necessarily labeled)}
}
\description{
Two-mode triad census to incremental triad census
}

