% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tmtc2setc}
\alias{tmtc2setc}
\title{Two-mode triad census}
\usage{
tmtc2setc(tmtc)
}
\arguments{
\item{bigraph}{The ambient affiliation network}

\item{graph}{A one-mode network (used as the one-mode projection of bigraph)}

\item{type}{The actor node type in bigraph}

\item{rcnames}{Logical; whether to label the matrix rows and columns}

\item{verbose}{Logical; whether to display progress bars}

\item{tmtc}{A full two-mode triad census (not necessarily labeled)}
}
\description{
These functions compute triad censuses for affiliation networks. (Currently
only the full triad census is implemented and the structural and simple
censuses are computed from it.) The full and structural censuses return
matrices, the former of whatever dimension is needed to contain the census
and the latter 4-by-2. The simple census returns a vector of length 4.
}

