% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tmtc2utc}
\alias{tmtc2utc}
\title{Two-mode triad census to uniformity triad census}
\usage{
tmtc2utc(tmtc)
}
\arguments{
\item{tmtc}{A full two-mode triad census (not necessarily labeled)}
}
\description{
Two-mode triad census to uniformity triad census
}

