% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transitivity.an.R
\name{transitivity.an}
\alias{transitivity.an}
\title{Affiliation network clustering coefficients}
\usage{
transitivity.an(bigraph, type = "global", wedgeFun = injequ.wedges,
  vids = which(!V(bigraph)$type), add.names = FALSE)
}
\arguments{
\item{bigraph}{An affiliation network; see \code{is.an}.}

\item{type}{Character; the type of clustering coefficient (defaults to
"global").}

\item{wedgeFun}{The wedge function.}

\item{vids}{A subset of actor node ids at which to evaluate the local
clustering coefficient.}

\item{add.names}{Logical; whether to label the matrix rows and columns.}
}
\value{
If \code{type} is "global", the global clustering coefficient of the
network; if "local", the local clustering coefficients of the actors;
otherwise, a 2-column matrix, each row of which gives the number of wedges
and the number of closed wedges centered at each actor.
}
\description{
This function computes a given flavor of transitivity (triadic closure) on a
given affiliation network. The calculations are performed locally. Each
flavor is defined as a proportion of "wedges" that are "closed", for suitable
definitions of both terms. The function `transitivity.an` is a shell that
proceeds across actors and computes wedges using the provided `wedgeFun`.
These functions count the "wedges", and among them the "closed" ones,
centered at a given actor node in a given affiliation network.
}
\examples{
data(women.clique)
sapply(c(injequ.wedges, injstr.wedges, indstr.wedges),
         transitivity.an, bigraph = women.clique, type = "local")
}

