% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transitivity.an.triads.R
\name{transitivity.an.triads}
\alias{transitivity.an.triads}
\title{Affiliation network clustering coefficients: triads method}
\usage{
transitivity.an.triads(bigraph, vids = which(!V(bigraph)$type), flavor)
}
\arguments{
\item{bigraph}{An affiliation network; see \code{is.an}.}

\item{vids}{A subset of actor node ids at which to evaluate the local
clustering coefficient.}

\item{flavor}{The flavor of transitivity.}
}
\value{
A two-column matrix of wedge counts and closed wedge counts at the nodes in vids.
}
\description{
This method first classifies every triad centered at each node. The appropriate formula then counts the wedges and closed wedges at each. The method is slower for a single flavor but can be used to produce multiple flavors with negligible additional computational cost.
}

