% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transitivity.an.wedges.R
\name{transitivity.an.wedges}
\alias{transitivity.an.wedges}
\title{Affiliation network clustering coefficients: wedges method}
\usage{
transitivity.an.wedges(bigraph, vids = which(!V(bigraph)$type), wedgeFun)
}
\arguments{
\item{bigraph}{An affiliation network; see \code{is.an}.}

\item{vids}{A subset of actor node ids at which to evaluate the local
clustering coefficient.}

\item{wedgeFun}{The wedge function.}
}
\value{
A two-column matrix of wedge counts and closed wedge counts at the nodes in vids.
}
\description{
This method relies on a separate "wedge function" for each statistic. The algorithm calls the appropriate wedge function to run over the necessary wedge centers and return a wedge count matrix, which is returned back into \code{\link{transitivity.an}} for outputting.
}

