% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad-closure-projection.r
\name{triad-closure-projection}
\alias{triad-closure-projection}
\alias{triad_closure_projection}
\title{Classical triad closure (on projections) for affiliation networks}
\usage{
triad_closure_projection(bigraph, actors = V(bigraph)[V(bigraph)$type ==
  FALSE], type = "global")
}
\arguments{
\item{bigraph}{An affiliation network.}

\item{actors}{A vector of actor nodes in \code{bigraph}.}

\item{type}{The type of statistic, matched to \code{"global"}, 
\code{"local"}, or \code{"raw"}.}
}
\value{
If \code{type} is \code{"global"}, the global statistic for 
  \code{bigraph}; if \code{"local"}, the local statistics for \code{actors}; 
  if \code{"raw"}, a 2-column matrix, each row of which gives the number of
  wedges and of closed wedges centered at \code{actors}.
}
\description{
Given an affiliation network and a vector of actor node IDs, 
  calculate the classical triad closure (clustering coefficient) centered at 
  the actors in the unipartite projection onto the actor nodes.
}
\details{
This function recapitulates
  \code{\link{triad_closure_watts_strogatz}} by invoking the
  \code{\link[igraph]{bipartite_projection}} and
  \code{\link[igraph]{transitivity}} functions in \strong{igraph}.
}
\seealso{
\code{\link{triad_closure}}

Other triad closure: \code{\link{triad-closure}}
}
