% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/triad.census.an.R
\name{triad.census.an}
\alias{triad.census.an}
\title{Affiliation network (full) triad census}
\usage{
triad.census.an(bigraph, add.names = FALSE, verbose = FALSE)
}
\arguments{
\item{bigraph}{An affiliation network.}

\item{add.names}{Logical; whether to label the matrix rows and columns}

\item{verbose}{Logical; whether to display progress bars}
}
\description{
This function computes the full triad census for an affiliation network.
}
\examples{
data(women.clique)
tc <- triad.census.an(women.clique, add.names = TRUE)
tc
sum(tc) == choose(vcount(actor.projection(women.clique)), 3)
}

