% Generated by roxygen2 (4.0.1): do not edit by hand
\name{triad.class}
\alias{triad.class}
\title{Triad classification}
\usage{
triad.class(bigraph, v)
}
\arguments{
\item{bigraph}{An affiliation network}

\item{v}{A subset of three actor nodes in bigraph}
}
\description{
Given three actor nodes in an affiliation network, determine the isomorphism
class of the triad they form by its index (a partition of three parts and a
nonnegative integer).
}

