% Generated by roxygen2 (4.0.1): do not edit by hand
\name{triad.plot}
\alias{triad.plot}
\title{Two-mode triads}
\usage{
triad.plot(lambda, w, layout, scale = 0.3, angdir = -1, rot = -pi/2,
  rot.lambda = c(0, 0, 0), rot.w = pi/12, ...)
}
\arguments{
\item{lambda}{The vector of pairwise weights}

\item{w}{The triadwise weight}
}
\description{
Two-mode triads
}

