% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/an-triad.R
\name{triad}
\alias{triad}
\alias{is_triad}
\alias{an_triad}
\alias{layout_triad}
\alias{plot_triad}
\alias{is.triad}
\alias{an.triad}
\alias{layout.triad}
\alias{plotTriad}
\title{Affiliation network triads}
\usage{
is_triad(graph)

an_triad(lambda, w, actors = letters[16:18], events = if (sum(c(lambda, w))
  == 0) c() else as.character(1:sum(c(lambda, w))))

layout_triad(lambda, w, scale = 0.3, angdir = -1, rot = -pi/2,
  rot_lambda = c(0, 0, 0), rot_w = pi/12)

plot_triad(lambda, w, layout, scale = 0.3, cex = 1, angdir = -1,
  rot = -pi/2, rot_lambda = c(0, 0, 0), rot_w = pi/12,
  actors = letters[16:18], events = if (sum(c(lambda, w)) == 0) c() else
  as.character(1:sum(c(lambda, w))), xlim, ylim, ...)

is.triad(graph)

an.triad(lambda, w, actors = letters[16:18], events = if (sum(c(lambda, w))
  == 0) c() else as.character(1:sum(c(lambda, w))))

layout.triad(lambda, w, scale = 0.3, angdir = -1, rot = -pi/2,
  rot_lambda = c(0, 0, 0), rot_w = pi/12)

plotTriad(lambda, w, layout, scale = 0.3, cex = 1, angdir = -1,
  rot = -pi/2, rot_lambda = c(0, 0, 0), rot_w = pi/12,
  actors = letters[16:18], events = if (sum(c(lambda, w)) == 0) c() else
  as.character(1:sum(c(lambda, w))), xlim, ylim, ...)
}
\arguments{
\item{graph}{An affiliation network; see \code{\link{is_an}}.}

\item{lambda}{The vector of pairwise weights}

\item{w}{The triadwise weight}

\item{actors}{Actor names (defaults to "P", "Q", and "R")}

\item{events}{Event names (defaults to positive integers)}

\item{scale}{A scaling parameter}

\item{angdir}{A rotation direction parameter (-1 for clockwise, 1 for
counter-clockwise)}

\item{rot}{An orientation parameter for the entire triad}

\item{rot_lambda}{An orientation parameter specifically for exclusive events}

\item{rot_w}{An orientation parameter specifically for inclusive events}

\item{layout}{Two-column matrix interpretable as an igraph layout}

\item{cex}{Node size scaling parameter}

\item{xlim}{Bounds on the horizontal axis}

\item{ylim}{Bounds on the vertical axis}

\item{...}{Additional arguments passed to \code{\link{plot}}.}
}
\description{
Affiliation network triads
}
