% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad-census-an.R
\name{triad_census_an}
\alias{triad_census_an}
\alias{triad.census.an}
\title{Affiliation network (full) triad census}
\usage{
triad_census_an(bigraph, add.names = FALSE, verbose = FALSE)

triad.census.an(bigraph, add.names = FALSE, verbose = FALSE)
}
\arguments{
\item{bigraph}{An affiliation network.}

\item{add.names}{Logical; whether to label the matrix rows and columns}

\item{verbose}{Logical; whether to display progress bars}
}
\description{
This function computes the full triad census for an affiliation network.
}
\examples{
data(women_clique)
tc <- triad_census_an(women_clique, add.names = TRUE)
tc
sum(tc) == choose(vcount(actor_projection(women_clique)), 3)
}
