% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad-class.R
\name{triad_class}
\alias{triad_class}
\title{Triad classification}
\usage{
triad_class(bigraph, v, sorted = TRUE)
}
\arguments{
\item{bigraph}{An affiliation network}

\item{v}{A subset of three actor nodes in bigraph}

\item{sorted}{Whether to sort the exclusive events, versus reporting them in
order of the nodes; defaults to TRUE}
}
\description{
Given three actor nodes in an affiliation network, determine the isomorphism 
class of the triad they form by its index (a partition of three parts and a 
nonnegative integer).
}

