% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/triadWedges.R
\name{triadWedges}
\alias{triadWedges}
\title{Wedge counts from triad classes}
\usage{
triadWedges(triads, flavor)
}
\arguments{
\item{triads}{A matrix of centered triads; see \code{link{centeredTriads}}.}

\item{flavor}{The flavor of transitivity to be used.}
}
\value{
A two-column matrix of wedge counts and closed wedge counts of flavor \code{flavor}.
}
\description{
Wedge counts from triad classes
}

