% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/two.tied.triads.R
\name{two.tied.triads}
\alias{two.tied.triads}
\title{Triad tallies}
\usage{
two.tied.triads(graph)
}
\arguments{
\item{graph}{A one-mode network}

\item{bigraph}{The ambient affiliation network from which graph is projected}
}
\description{
These functions are called by the full triad census to handle triads of
different types using the projection onto actor nodes. The name of each
function indicates the number of edges that appear among the three actors of
the triad in the projection. (Zero-edge triads do not need to be tallied;
the total number of triads is easily calculated, and the difference between
this number and the total number of triads with edges gives the number of
triads without.)
}

