% Generated by roxygen2 (4.0.1): do not edit by hand
\name{twomode.triad}
\alias{twomode.triad}
\title{Two-mode triads}
\usage{
twomode.triad(lambda, w, actors = LETTERS[16:18], events = if (sum(c(lambda,
  w)) == 0) c() else as.character(1:sum(c(lambda, w))))
}
\arguments{
\item{lambda}{The vector of pairwise weights}

\item{w}{The triadwise weight}

\item{actors}{Actor names (defaults to 'P', 'Q', and 'R')}

\item{events}{Event names (defaults to positive integers)}
}
\description{
Two-mode triads
}

