% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unif-triad-census.R
\name{unif_triad_census}
\alias{unif_triad_census}
\alias{unif.triad.census}
\title{Uniformity triad census}
\usage{
unif_triad_census(bigraph)

unif.triad.census(bigraph)
}
\arguments{
\item{bigraph}{The affiliation network}
}
\description{
This function computes the uniformity triad census for an affiliation
network, an 8-by-2 array of frequency counts for each isomorphism class of
triad, modulo the equivalence relation that identifies triads based on the
positivity of their triad weights w and of the differences between their
sequential pairwise weights. This census can be recovered from the full triad
census, and the structural equivalence triad census can be recovered from it.
The census must sum to choose(n, 3), where n is
the number of actor nodes.
}
\details{
\strong{This function is deprecated and may be removed in the next version.}
}
