% Generated by roxygen2 (4.0.1): do not edit by hand
\name{utc2ctc}
\alias{utc2ctc}
\title{Uniformity triad census to cooperativity triad census}
\usage{
utc2ctc(utc)
}
\arguments{
\item{utc}{A uniformity two-mode triad census (not necessarily labeled)}
}
\description{
Uniformity triad census to cooperativity triad census
}

