% Generated by roxygen2 (4.0.1): do not edit by hand
\name{utc2stc}
\alias{utc2stc}
\title{Uniformity triad census to simple triad census}
\usage{
utc2stc(utc)
}
\arguments{
\item{utc}{A uniformity two-mode triad census (not necessarily labeled)}
}
\description{
Uniformity triad census to simple triad census
}

