% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc2tc.R
\name{utc2tc}
\alias{utc2tc}
\title{Uniformity triad census to simple triad census}
\usage{
utc2tc(tc)
}
\arguments{
\item{tc}{A uniformity triad census (not necessarily labeled)}
}
\description{
TO BE DELETED
}

