% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wedgecount-census.R
\name{wedgecount_census}
\alias{wedgecount_census}
\title{Global clustering coefficients from the full triad census}
\usage{
wedgecount_census(census, closedFun, openFun)
}
\arguments{
\item{census}{Numeric matrix; a full triad census}

\item{closedFun}{The closed wedge count for a triad}

\item{openFun}{The open wedge count for a triad}
}
\description{
Each global clustering coefficient can be recovered from the full triad
census. the first function provides a framework for this calculation, and the
following call that framework for specific clustering coefficients.
}

