% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wedges-an-flavors.R, R/wedges-an.r
\name{indequ_wedges}
\alias{indequ_wedges}
\alias{indequ.wedges}
\alias{indstr_wedges}
\alias{indstr.wedges}
\alias{injact_wedges}
\alias{injact.wedges}
\alias{injequ_wedges}
\alias{injequ.wedges}
\alias{injstr_wedges}
\alias{injstr.wedges}
\alias{wedges_an}
\title{Affiliation network wedges}
\usage{
indequ_wedges(bigraph, Q)

indequ.wedges(bigraph, Q)

indstr_wedges(bigraph, Q)

indstr.wedges(bigraph, Q)

injact_wedges(bigraph, Q)

injact.wedges(bigraph, Q)

injequ_wedges(bigraph, Q)

injequ.wedges(bigraph, Q)

injstr_wedges(bigraph, Q)

injstr.wedges(bigraph, Q)

wedges_an(bigraph, q, alcove = 0, wedge = 0, maps = 0, congruence = 0)
}
\arguments{
\item{bigraph}{An affiliation network.}

\item{Q}{An actor node in the network.}

\item{q}{An actor node in \code{bigraph}.}

\item{alcove, wedge, maps, congruence}{Choice of alcove, wedge, maps, and
congruence (see Details).}
}
\description{
Each clustering coefficient can be defined as the proportion of "wedges" that
are "closed", for suitable definitions of both terms. These functions count
the "wedges", and among them the "closed" ones, centered at a given actor
node in a given affiliation network.

Given an affiliation network and a vector of node IDs, list all 
  wedges of a specified type centered at each node and indicate whether each 
  is closed.
}
\details{
Each measure of transitive closure implemented here is defined as 
  the proportion of wedges that are closed, where a \emph{wedge} is the image
  of a specified self-dual two-event triad \eqn{W} under a specified 
  subcategory of graph maps \eqn{\mathcal{C}} subject to a specified 
  congruence relation \eqn{\tilde}, and where a wedge is \emph{closed} if it 
  is the image of such a map that factors through a canonical inclusion of 
  \eqn{W} to a specified self-dual three-event triad \eqn{X}.
  
  The alcove, wedge, maps, and congruence can be specified by numerical codes
  as follows:
  \itemize{
   \item{"alcove:"}{
    \itemize{0}{\eqn{T_{(1,1,1),0}}}
    \itemize{1}{\eqn{T_{(1,1,0),1}}}
    \itemize{2}{\eqn{T_{(1,0,0),2}}}
    \itemize{3}{\eqn{T_{(0,0,0),3}}}
   }
   \item{"wedge:"}{
    \itemize{0}{\eqn{T_{(1,1,0),0}}}
    \itemize{1}{\eqn{T_{(1,0,0),1}}}
    \itemize{2}{\eqn{T_{(0,0,0),2}}}
   }
   \item{"maps:"}{
    \itemize{0}{all graph maps (injective on actors)}
    \itemize{1}{injective graph maps}
    \itemize{2}{induced injective graph maps}
   }
   \item{"congruence:"}{
    \itemize{0}{same actor images}
    \itemize{1}{same actor images, structurally equivalent event images}
    \itemize{2}{same actor images, same event images}
   }
  }

See the references for a full definition.
}
\references{
Brunson, J.C. (2015). Triadic analysis of affiliation networks. 
  *Network Science*, 3 (4), 480--508.
}
