% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wobble.coords}
\alias{wobble.coords}
\title{Two-mode triads}
\usage{
wobble.coords(l, S = 0.001)
}
\arguments{
\item{l}{Layout matrix, having two columns}

\item{S}{Wobble coefficient (should be small, defaults to .001)}
}
\description{
Two-mode triads
}

