% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{women_clique}
\alias{women_clique}
\title{Clique of women connected by event coattendance in Old City.}
\format{An affiliation network; see \code{\link{is_an}}.}
\source{
Davis, Gardner, and Gardner (1941) *Deep South: A Social
Anthropological Study of Caste and Class*
\url{http://books.google.com/books?id=HGIdAAAAIAAJ}
}
\description{
These data record the participation of 5 African-American women in 5 social
activities.
}

