% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/actor.count.R
\name{actor.count}
\alias{actor.count}
\title{Count the actors in an affiliation network}
\usage{
actor.count(bigraph)
}
\arguments{
\item{bigraph}{An affiliation network; see `is.an`.}
}
\description{
This function returns the number of actors (nodes of type FALSE) in an
affiliation network.
}
\examples{
data(southafrica1905)
actor.count(southafrica1905)
}

