% Generated by roxygen2 (4.0.1): do not edit by hand
\name{actor.projection}
\alias{actor.projection}
\title{Projection onto actors of an affiliation network}
\usage{
actor.projection(bigraph, type = 0, name = "name")
}
\arguments{
\item{bigraph}{The ambient affiliation network}

\item{type}{The node type (0 or 1) of the actors of bigraph}

\item{name}{What attribute of the actor nodes in bigraph to name the nodes
in the projection (defaults to 'name')}
}
\description{
This function uses the igraph function bipartite.projection to compute the
projection of an affiliation network onto the actor nodes.
}

