% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/an.R
\name{an}
\alias{an}
\alias{as.an}
\alias{as_an}
\alias{is.an}
\alias{is_an}
\title{Affiliation network}
\usage{
is_an(graph)

as_an(graph)

is.an(graph)

as.an(graph)
}
\arguments{
\item{graph}{An igraph object}
}
\description{
An affiliation network is, essentially, a bipartite graph. The igraph 
function \code{\link{is_bipartite}} tests an igraph object for the logical
vertex attribute \code{type} (if it has any vertices at all), which is
intended to provide a bipartition of the vertices. It does not, however, test
whether edges exist within either part of the partition. The function
\code{\link{is_an}} tests this condition. To simplify some functions,
affiliation networks, as defined here, must also have their vertices indexed
in order of part; all actor vertices must precede all event vertices. The
function \code{is_an} tests also for this condition. The coersive function
\code{as_an} imposes, by a minimal permutation of vertex labels, provided the
\code{igraph} object is bipartite to begin with.
}

