% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/an.triad.census.R
\name{an.triad.census}
\alias{an.triad.census}
\title{Affiliation network (full) triad census}
\usage{
an.triad.census(bigraph, type = 0, rcnames = FALSE, verbose = FALSE)
}
\arguments{
\item{bigraph}{The ambient affiliation network}

\item{type}{The actor node type in bigraph}

\item{rcnames}{Logical; whether to label the matrix rows and columns}

\item{verbose}{Logical; whether to display progress bars}
}
\description{
These functions compute triad censuses for affiliation networks. (Currently
only the full triad census is implemented and the structural and simple
censuses are computed from it.) The full and structural censuses return
matrices, the former of whatever dimension is needed to contain the census
and the latter 4-by-2. The simple census returns a vector of length 4.
}

