% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/an.triad.census.R
\name{an.triad.census}
\alias{an.triad.census}
\title{Affiliation network (full) triad census}
\usage{
an.triad.census(bigraph, type = 0, rcnames = FALSE, verbose = FALSE)
}
\arguments{
\item{bigraph}{An affiliation network.}

\item{type}{Logical; the node type in `bigraph` from which triads are taken.}

\item{rcnames}{Logical; whether to label the matrix rows and columns}

\item{verbose}{Logical; whether to display progress bars}
}
\description{
This function computes the full triad census for an affiliation network.
}

