% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{chicago1960s}
\alias{chicago1960s}
\title{Corporate interlocks among directors in Chicago.}
\format{An affiliation network; see \code{\link{is.an}}.}
\source{
Barnes & Burkett (2010) "Structural Redundancy and Multiplicity in
Corporate Networks", *Connections* **30**(2), p. 4-20.
\url{http://www.insna.org/PDF/Connections/v30/2010_I-2_P-1-1.pdf}
}
\description{
An affiliation network of 20 directors of 24 corporations and social clubs.
}

