% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/combinatorial-bijections.R
\name{index_subset}
\alias{index_subset}
\alias{subset_index}
\alias{subset_partition}
\alias{partition_subset}
\alias{index_partition}
\alias{partition_index}
\alias{combinatorial_bijections}
\alias{indexSubset}
\alias{indexPartition}
\alias{subsetIndex}
\alias{subsetPartition}
\alias{partitionIndex}
\alias{partitionSubset}
\title{Combinatorial bijections for affiliation network triad indexing}
\usage{
index_subset(i)

subset_index(v)

subset_partition(v)

partition_subset(par)

index_partition(i)

partition_index(par)

indexSubset(i)

indexPartition(i)

subsetIndex(v)

subsetPartition(v)

partitionIndex(par)

partitionSubset(par)
}
\arguments{
\item{i}{Integer; an index in the total order, starting at 0.}

\item{v}{Integer vector; a set of 3 distinct non-negative integers, in 
decreasing order.}

\item{par}{Integer vector; a partition of at most 3 parts, with parts in 
non-increasing order.}
}
\description{
These functions biject among partitions of at most 3 parts, 
  3-subsets of natural numbers, and indices for the lexicographic total 
  orders on them.
}
\examples{
index_subset(2)
index_partition(2)
subset_index(c(3, 2, 0))
subset_partition(c(3, 2, 0))
partition_index(c(1, 1, 0))
partition_subset(c(1, 1, 0))
}
