% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorial_bijections.R
\name{combinatorial_bijections}
\alias{combinatorial_bijections}
\alias{index_partition}
\alias{index_subset}
\alias{partition_index}
\alias{partition_subset}
\alias{subset_index}
\alias{subset_partition}
\title{Combinatorial bijections for affiliation network triad indexing}
\usage{
index_subset(i)

index_partition(i)

subset_index(vec)

subset_partition(vec)

partition_index(lambda)

partition_subset(lambda)
}
\arguments{
\item{i}{An index in the total order; starts at 0.}

\item{vec}{A (decreasing) 3-subset of natural numbers (including 0).}

\item{lambda}{A partition of at most 3 parts.}
}
\description{
These functions biject among partitions of at most 3 parts, 3-subsets of
natural numbers, and indices for the lexicographic total orders on them.
}
\examples{
index_subset(2)
index_partition(2)
subset_index(c(3, 2, 0))
subset_partition(c(3, 2, 0))
partition_index(c(1, 1, 0))
partition_subset(c(1, 1, 0))
}

