% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dual.an.R
\name{dual.an}
\alias{dual.an}
\title{Take the dual of an affiliation network}
\usage{
dual.an(bigraph)
}
\arguments{
\item{bigraph}{An affiliation network; see \code{\link{is.an}}.}
}
\description{
This function negates the logical values of the vertex \code{type} attribute
and reorders the vertex ids accordingly, effectively obtaining the dual
affiliation network.
}
\examples{
data(southafrica1905)
tab <- table(V(southafrica1905)$type)
proj <- actor.projection(dual.an(southafrica1905))
vcount(proj) == tab[2]
}

