% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualize.r
\name{dual}
\alias{dual}
\alias{dualize}
\alias{dual_an}
\alias{dual.an}
\title{Take the dual of an affiliation network}
\usage{
dualize(bigraph)

dual_an(bigraph)

dual.an(bigraph)
}
\arguments{
\item{bigraph}{An affiliation network.}
}
\description{
This function obtains the dual of an affiliation network, in
  which the actors and events have swapped roles. To do this, it negates
  the logical values of the node \code{type} attribute and reorders the
  node ids accordingly.
}
\examples{
data(women_clique)
tab <- table(V(women_clique)$type)
proj <- actor_projection(dual_an(women_clique))
vcount(proj) == tab[2]
}
