% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-an.R
\name{dynamic_an}
\alias{dynamic_an}
\alias{is_dynamic_an}
\alias{is.dyn}
\title{Dynamic graphs}
\usage{
is_dynamic_an(graph)

is.dyn(graph)
}
\arguments{
\item{graph}{An \code{igraph} object.}
}
\description{
An affiliation network is \emph{dynamic}, for present purposes, 
  if its event nodes have time stamps, recorded as a numeric vertex attribute
  \code{"time"}.
}
