% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-an.R
\name{dynamic_an}
\alias{dynamic_an}
\alias{is_dynamic_an}
\alias{is.dyn}
\alias{as_dynamic_an}
\title{Dynamic affiliation network structure}
\usage{
is_dynamic_an(graph)

is.dyn(graph)

as_dynamic_an(graph, use.time.attribute = "time",
  add.time.attribute = FALSE)
}
\arguments{
\item{graph}{An affiliation network.}

\item{use.time.attribute}{Character; the vertex attribute to coerce to 
numeric if necessary and use as the \code{"time"} attribute.}

\item{add.time.attribute}{Logical; whether, if \code{graph} has no attribute
\code{use.time.attribute}, to introduce an artificial \code{"time"}
attribute taking the values \code{1:event_count(graph)}, reflecting the
order of the event node IDs.}
}
\description{
An affiliation network is \emph{dynamic}, for present purposes, 
  if its event nodes have time stamps, recorded as a numeric vertex attribute
  \code{"time"} in non-decreasing order.
}
\seealso{
Other network testing and coercion: \code{\link{affiliation_network}}
}
