% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-an.R
\name{dynamic_an}
\alias{dynamic_an}
\alias{is_dynamic_an}
\alias{is.dyn}
\alias{as_dynamic_an}
\title{Dynamic affiliation network structure}
\usage{
is_dynamic_an(graph)

is.dyn(graph)

as_dynamic_an(graph, add.time.attribute = FALSE)
}
\arguments{
\item{graph}{An affiliation network.}

\item{add.time.attribute}{Character or logical; if character, the existing 
attribute to coerce to numeric if necessary and use as the time attribute;
if logical, whether to introduce an artificial \code{time} attribute
\code{1:event_count(graph)}, reflecting the order of the event node IDs,
if \code{graph} has none.}
}
\description{
An affiliation network is \emph{dynamic}, for present purposes, 
  if its event nodes have time stamps, recorded as a numeric vertex attribute
  \code{"time"} in non-decreasing order.
}
\seealso{
Other network testing and coercion: \code{\link{affiliation_network}}
}
