% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-dynamic-an.R
\name{dynamic_an}
\alias{dynamic_an}
\alias{is.dyn}
\alias{is_dynamic_an}
\title{Dynamic graphs}
\usage{
is_dynamic_an(graph)

is.dyn(graph)
}
\arguments{
\item{graph}{An igraph object.}
}
\description{
A simple graph is dynamic if its edges have time stamps. A bipartite graph is
dynamic if its event nodes have time stamps.
}

