% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-triad-closure.R
\name{dynamic_triad_closure}
\alias{dynamic_triad_closure}
\alias{dynamic_triad_closure_an}
\alias{dynamic_transitivity_an}
\alias{dyn.transitivity.an}
\alias{dynamic_triad_closure_projection}
\title{Triadic closure for dynamic affiliation networks}
\usage{
dynamic_triad_closure(graph, actors = V(graph)[V(graph)$type == FALSE],
  type = "global", ..., measure = NULL)

dynamic_triad_closure_an(graph, actors = V(graph)[V(graph)$type == FALSE],
  type = "global", ..., measure = NULL)

dynamic_transitivity_an(graph, actors = V(graph)[V(graph)$type == FALSE],
  type = "global", ..., measure = NULL)

dyn.transitivity.an(graph, actors = V(graph)[V(graph)$type == FALSE],
  type = "global", ..., measure = NULL)

dynamic_triad_closure_projection(graph, memory = Inf, type = "global")
}
\arguments{
\item{graph}{An affiliation network with time-stamped events.}

\item{actors}{A vector of actor nodes in \code{graph}.}

\item{type}{The type of statistic, matched to \code{"global"}, 
\code{"local"}, or \code{"raw"}.}

\item{...}{Additional parameters passed to specific functions.}

\item{measure}{Character; the measure of triad closure, used as the suffix 
\code{*} to \code{triad_closure_*}. Matched to \code{"classical"} (also 
\code{"watts_strogatz"}), \code{"twomode"} (also \code{"opsahl"}), 
\code{"unconnected"} (also \code{"liebig_rao_0"}), 
\code{"completely_connected"} (also \code{"liebig_rao_3"}), 
\code{"exclusive"}, or \code{"projection"}.}

\item{memory}{Numeric; minimum delay of wedge formation since would-have-been
closing events.}
}
\description{
Given an affiliation network with time-stamped events, compute the proportion
of centered triples at which an open wedge exists at some time that is closed
at a later time.
}
\examples{
data(women_group)
dynamic_triad_closure(women_group)
cbind(
  transitivity(actor_projection(women_group), type = "local"),
  triad_closure_opsahl(women_group, type = "local"),
  triad_closure_exclusive(women_group, type = "local"),
  dynamic_triad_closure_projection(women_group, type = "local"),
  dynamic_triad_closure(women_group, type = "local")
)
}
\seealso{
Other triad closure functions: \code{\link{project_transitivity}},
  \code{\link{transitivity_an}},
  \code{\link{triad_closure_from_census}},
  \code{\link{triad_closure}}
}
