% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/event.count.R
\name{event.count}
\alias{event.count}
\title{Number of events in an affiliation network}
\usage{
event.count(bigraph)
}
\arguments{
\item{bigraph}{The ambient affiliation network}
}
\description{
This function returns the number of events (nodes of type TRUE) in an
affiliation network.
}
\examples{
data(inner.circle)
event.count(inner.circle)
}

