% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/event.projection.R
\name{event.projection}
\alias{event.projection}
\title{Projection onto events of an affiliation network}
\usage{
event.projection(bigraph, name = "name")
}
\arguments{
\item{bigraph}{The ambient affiliation network}

\item{name}{What attribute of the event nodes in bigraph to name the nodes
in the projection (defaults to 'name')}
}
\description{
This function uses the igraph function bipartite.projection to compute the
projection of an affiliation network onto the event nodes.
}
\examples{
data(hobson.inner.circle)
tab <- table(V(hobson.inner.circle)$type)
proj <- event.projection(hobson.inner.circle)
vcount(proj) == tab[2]
}

