% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ftc2injequ.R
\name{ftc2injequ}
\alias{ftc2injequ}
\title{Global clustering coefficients from the triad census}
\usage{
ftc2injequ(census)
}
\arguments{
\item{census}{Numeric matrix; a full triad census}
}
\description{
Global clustering coefficients from the triad census
}

