% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hall.criterion}
\alias{hall.criterion}
\title{Hall criterion}
\usage{
hall.criterion(lst)
}
\arguments{
\item{lst}{A list of vectors possibly having some common elements}
}
\description{
A criterion for the existence of a system of distinct representatives for a
finite set:
www.encyclopediaofmath.org/index.php/System_of_different_representatives
}

