% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hallCriterion.R
\name{hallCriterion}
\alias{hallCriterion}
\title{Hall criterion}
\usage{
hallCriterion(lst)
}
\arguments{
\item{lst}{A list of vectors possibly having some common elements}
}
\description{
Test for the existence of a system of distinct representatives for a finite
set.
\url{www.encyclopediaofmath.org/index.php/System_of_different_representatives}
}

