% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/index.subset.R
\name{index.subset}
\alias{index.subset}
\title{Combinatorial bijections for affiliation network triad labeling}
\usage{
index.subset(i)
}
\arguments{
\item{i}{An index under the total order. Starts at 0.}
}
\description{
These functions biject among partitions of at most 3 parts, 3-subsets of
natural numbers, and indices for the lexicographic total orders on both.
}
\examples{
index.subset(2)
}

