% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/indexSubset.R
\name{indexSubset}
\alias{indexSubset}
\title{Combinatorial bijections for affiliation network triad indexing}
\usage{
indexSubset(i)
}
\arguments{
\item{i}{An index in the total order; starts at 0.}
}
\value{
A (decreasing) 3-subset of natural numbers (including 0).
}
\description{
These functions biject among partitions of at most 3 parts, 3-subsets of
natural numbers, and indices for the lexicographic total orders on them.
}
\examples{
indexSubset(2)
}

