% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/indstr.wedges.R
\name{indstr.wedges}
\alias{indstr.wedges}
\title{Affiliation network wedges}
\usage{
indstr.wedges(bigraph, Q)
}
\arguments{
\item{bigraph}{The ambient affiliation network.}

\item{Q}{An actor node in the network.}
}
\description{
Each clustering coefficient can be defined as the proportion of "wedges" that
are "closed", for suitable definitions of both terms. These functions count
the "wedges", and among them the "closed" ones, centered at a given actor
node in a given affiliation network.
}
\seealso{
Other wedge functions: \code{\link{indequ.wedges}};
  \code{\link{injact.wedges}}; \code{\link{injequ.wedges}};
  \code{\link{injstr.wedges}}
}

