% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode-addition.r
\name{mode_addition}
\alias{mode_addition}
\alias{add_actors}
\alias{add_events}
\title{Add actor and event nodes}
\usage{
add_actors(graph, nv, ..., attr = list(), events = NULL)

add_events(graph, nv, ..., attr = list(), actors = NULL)
}
\arguments{
\item{graph}{An affiliation network.}

\item{nv, ..., attr}{Arguments passed to \code{\link[igraph]{add_vertices}}. 
Events added to a dynamic affiliation network should be given time 
attributes.}

\item{events}{A vector of event nodes in \code{graph}, to be linked to the 
new actor(s).}

\item{actors}{A vector of actor nodes in \code{graph}, to be linked to the 
new event(s).}
}
\description{
These functions add actor and event nodes (as desired) to a 
  graph while maintaining its (temporal) affiliation network structure.
}
\examples{
data(women_clique)
plot(prettify_an(add_actors(women_clique, nv = 1, events = c(7, 9))))
data(women_group)
plot(prettify_an(women_group))
actor_names <- c("Frances", "Dorothy")
cbind(
  dynamic_triad_closure(women_group, type = "local"),
  dynamic_triad_closure(
    add_events(women_group, nv = 1, actors = actor_names, time = 0),
    type = "local"
  ),
  dynamic_triad_closure(
    add_events(women_group, nv = 1, actors = actor_names, time = 367),
    type = "local"
  )
)
}
\seealso{
Original \strong{igraph} functions:
  \code{\link[igraph]{add_vertices}}, \code{\link[igraph]{add_edges}}

Other modal queries and manipulations: \code{\link{dualize}},
  \code{\link{mode_counts}}, \code{\link{modes}},
  \code{\link{schedule}}
}
