% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode-counts.R
\name{mode_counts}
\alias{mode_counts}
\alias{actor_count}
\alias{event_count}
\alias{actor.count}
\alias{event.count}
\title{Count the actors and events in an affiliation network}
\usage{
actor_count(graph)

event_count(graph)

actor.count(graph)

event.count(graph)
}
\arguments{
\item{graph}{An affiliation network.}
}
\description{
These functions return the number of actors (nodes with 
  \code{type} attribute \code{FALSE}) or events (\code{TRUE}) in an 
  affiliation network.
}
\examples{
data(chicago1960s)
actor_count(chicago1960s)
event_count(chicago1960s)
}
\seealso{
Original \strong{igraph} functions: \code{\link[igraph]{vcount}},
  \code{\link[igraph]{ecount}}

Other modal queries and manipulations: \code{\link{dualize}},
  \code{\link{mode_addition}}, \code{\link{modes}},
  \code{\link{schedule}}
}
