% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modes.r
\name{modes}
\alias{modes}
\alias{V1}
\alias{V2}
\alias{set_actor_attr}
\alias{set_event_attr}
\alias{V1<-}
\alias{V2<-}
\title{Actor and event node iteration and attribute querying and assignment}
\usage{
V1(graph)

V2(graph)

set_actor_attr(graph, name, index = V1(graph), value)

set_event_attr(graph, name, index = V2(graph), value)

V1(x) <- value

V2(x) <- value
}
\arguments{
\item{graph}{An affiliation network.}

\item{name}{The name of the attribute to set.}

\item{index}{An optional node sequence to set the attributes of a subset of 
actor or event nodes.}

\item{value}{The new value of the attribute for all (or \code{index}) actor 
or event nodes.}

\item{x}{An affiliation network.}
}
\value{
\code{graph}, with the actor or event attribute added or set.
}
\description{
These functions return actor and event node lists.

x = an affiliation network
value = V1(x) or V2(x) with additional attributes:
  "name" = name of attribute being assigned
  "value" = values being assigned to attribute
}
\examples{
data(women_clique)
V1(women_clique)
V2(women_clique)
}
\seealso{
\code{\link[igraph]{V}}, \code{\link[igraph]{set_vertex_attr}}
}
