% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{nmt.meetings}
\alias{nmt.meetings}
\title{Noordin Top meeting attendance network.}
\format{An undirected igraph object with nodes of logical attribute `type`;
individuals correspond to nodes of type `FALSE`, meetings to nodes of type
`TRUE`.}
\source{
The Association of Religious Data Archives
\url{http://www.thearda.com/Archive/Files/Descriptions/TERRNET.asp}
}
\description{
An affiliation network of 26 individuals who attended 20 meetings
associated with Noordin Mohammad Top.
}

