% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{nmt.organizations}
\alias{nmt.organizations}
\title{Noordin Top organization membership network.}
\format{An affiliation network; see \code{\link{is.an}}.}
\source{
The Association of Religious Data Archives
\url{http://www.thearda.com/Archive/Files/Descriptions/TERRNET.asp}
}
\description{
These data record the attendance of 67 members of 32 organizations
associated with Noordin Mohammad Top.
}

