% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/partitionSubset.R
\name{partitionSubset}
\alias{partitionSubset}
\title{Combinatorial bijections for affiliation network triad indexing}
\usage{
partitionSubset(lambda)
}
\arguments{
\item{lambda}{A partition of 3 parts (possibly including 0).}
}
\value{
A (decreasing) 3-subset of natural numbers (including 0).
}
\description{
These functions biject among partitions of at most 3 parts, 3-subsets of
natural numbers, and indices for the lexicographic total orders on them.
}
\examples{
partitionSubset(c(1, 1, 0))
}

