% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.triad.R
\name{plot.triad}
\alias{plot.triad}
\title{Plot affiliation network triads}
\usage{
plot.triad(lambda, w, layout, scale = 0.3, cex = 1, angdir = -1,
  rot = -pi/2, rot.lambda = c(0, 0, 0), rot.w = pi/12,
  actors = letters[16:18], events = if (sum(c(lambda, w)) == 0) c() else
  as.character(1:sum(c(lambda, w))), xlim, ylim, ...)
}
\arguments{
\item{lambda}{The vector of pairwise weights}

\item{w}{The triadwise weight}

\item{layout}{A layout matrix for the vertices of the triad (if missing, a
standard layout is constructed)}

\item{scale}{A scaling parameter}

\item{angdir}{A rotation direction parameter (-1 for clockwise, 1 for
counter-clockwise)}

\item{rot}{An orientation parameter for the entire triad}

\item{rot.lambda}{An orientation parameter specifically for exclusive events}

\item{rot.w}{An orientation parameter specifically for inclusive events}
}
\description{
Plot affiliation network triads
}

