% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify-an.R
\name{prettify}
\alias{prettify}
\alias{prettify_an}
\alias{prettify.an}
\title{Convenient plotting aesthetics for affiliation networks}
\usage{
prettify_an(graph)

prettify.an(graph)
}
\arguments{
\item{graph}{An affiliation network.}
}
\description{
Given an affiliation network, assign the node and link
  aesthetics to values that produce a neater visualization through
  \code{\link[igraph]{plot.igraph}} than the \strong{igraph} defaults.
}
\examples{
library(igraph)
data(women_clique)
data(whigs)
for (g in list(women_clique, whigs)) {
    plot(prettify_an(g))
}
}
