% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/project.census.R
\name{project.census}
\alias{project.census}
\title{Project a triad census}
\usage{
project.census(census, scheme = NULL, add.names = FALSE)
}
\arguments{
\item{census}{Numeric; the input triad census. It is classified as
"uniformity", "structural", or "simple" if its dimensons are 8-by-2, 4-by-2,
or 4-by-1, respectively, or unless otherwise specified; otherwise it is
classified as "full".}

\item{scheme}{Character; the type of census (to be matched to "full",
"uniformity", "structural", or "simple")}
}
\description{
This function inputes a triad census of any scheme and returns a list of
triad censuses projected from it (not icluding itself).
}

