% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-census.R
\name{project_census}
\alias{project_census}
\alias{project.census}
\alias{difference_from_full_census}
\alias{ftc2utc}
\alias{binary_from_full_census}
\alias{ftc2stc}
\alias{simple_from_full_census}
\alias{ftc2tc}
\alias{binary_from_difference_census}
\alias{utc2stc}
\alias{simple_from_difference_census}
\alias{utc2tc}
\alias{simple_from_binary_census}
\alias{stc2tc}
\title{Project a higher-resolution triad census to a lower-resolution one}
\usage{
project_census(census, scheme = NULL, add.names = FALSE)

project.census(census, scheme = NULL, add.names = FALSE)

difference_from_full_census(census)

ftc2utc(census)

binary_from_full_census(census)

ftc2stc(census)

simple_from_full_census(census)

ftc2tc(census)

binary_from_difference_census(census)

utc2stc(census)

simple_from_difference_census(census)

utc2tc(census)

simple_from_binary_census(census)

stc2tc(census)
}
\arguments{
\item{census}{Numeric matrix or vector; an affiliation network triad census. 
It is treated as binary or simple if its dimensons are 4-by-2 or 4-by-1, 
respectively, unless otherwise specified by \code{scheme}; otherwise it is 
treated as full.}

\item{scheme}{Character; the type of triad census provided, matched to 
\code{"full"}, \code{"difference"} (also \code{"uniformity"}), 
\code{"binary"} (also \code{"structural"}), or \code{"simple"}.}

\item{add.names}{Logical; whether to label the rows and columns of the output
matrix.}
}
\description{
Given a triad census of any scheme, construct a triad census of 
  a coarser (strictly less informative) scheme.
}
\details{
This function inputes an affiliation network triad census of any 
  scheme and returns a list of triad censuses projected from it (not icluding
  itself). The schemes are, in order of resolution, \emph{full} (also called 
  the \emph{affiliation network triad census} without qualification), 
  \emph{difference}, \emph{binary}, and \emph{simple}. A final element of the
  output list is the total number of triads in the affiliation network. Each
  summary can be recovered from those before it, specifically by aggregating
  certain matrix entries to form a smaller matrix. The helper functions
  \code{*_from_*_census()} project a census of each scheme to one of each
  coarser scheme.


}
\section{Triad censuses}{
 Three triad censuses are implemented for affiliation 
  networks:
  \itemize{
  \item The \emph{full triad census} (Brunson, 2015) records the number of 
  triads of each isomorphism class. The classes are indexed by a partition, 
  \eqn{\lambda=(\lambda_1\leq\lambda_2\leq\lambda_3)}, indicating the number 
  of events attended by both actors in each pair but not the third, and a 
  positive integer, \eqn{w}, indicating the number of events attended by all 
  three actors. The isomorphism classes are organized into a matrix with rows
  indexed by \eqn{\lambda} and columns indexed by \eqn{w}, with the 
  partitions \eqn{\lambda} ordered according to the \emph{revolving door 
  ordering} (Kreher & Stinson, 1999). The main function 
  \code{\link{triad_census_an}} (called from \code{triad_census} when the 
  \code{graph} argument is an \code{affiliation_network}) defaults to this
  census.
  \item For the analysis of sparse affiliation networks, the full triad 
  census may be less useful than information on whether the extent of 
  connectivity through co-attended events differs between each pair of 
  actors. In order to summarize this information, a coarser triad census can 
  be conducted on classes of triads based on the following congruence 
  relation: Using the indices \eqn{\lambda=(x\ge y\ge z)} and \eqn{w} 
  above, note that the numbers of shared events for each pair and for the 
  triad are \eqn{x+w\ge y+w\ge z+w\ge w\ge 0}. Consider two triads 
  congruent if the same subset of these weak inequalities are strictly 
  satisfied. The resulting \emph{difference triad census}, previously called 
  the \emph{uniformity triad census}, implemented as
  \code{\link{triad_census_difference}}, is organized into a \eqn{8\times 2} 
  matrix with the strictness of the first three inequalities determining the 
  row and that of the last inequality determining the column.
  \item A still coarser congruence relation can be used to tally how many are
  connected by at least one event in each distinct way. This relation 
  considers two triads congruent if each corresponding pair of actors both 
  attended or did not attend at least one event not attended by the third, 
  and if the corresponding triads both attended or did not attend at least 
  one event together. The \emph{binary triad census} (Brunson, 2015; therein 
  called the \emph{structural triad census}), implemented as 
  \code{\link{triad_census_binary}}, records the number of triads in each 
  congruence class.
  \item The \emph{simple triad census} is the 4-entry triad census on a 
  traditional (non-affiliation) network indicating the number of triads of 
  each isomorphism class, namely whether the triad contains zero, one, two, 
  or three links. The function \code{\link{simple_triad_census}} computes the
  classical (undirected) triad census for an undirected traditional network, 
  or for the actor projection of an affiliation network (if provided), using 
  \code{\link[igraph]{triad_census}}; if the result doesn't make sense (i.e.,
  the sum of the entries is not the number of triples of nodes), then it 
  instead uses its own, much slower method.
  }
  Each of these censuses can be projected from the previous using the 
  function \code{\link{project_census}}. A fourth census, called the 
  \emph{uniformity triad census} and implemented as 
  \code{\link{unif_triad_census}}, is deprecated. Three-actor triad 
  affiliation networks can be constructed and plotted using the 
  \code{\link{triad}} functions.
  
  The default method for the two affiliation network--specific triad censuses
  is adapted from the algorithm of Batagelj and Mrvar (2001) for calculating
  the classical triad census for a directed graph.
}

\references{
Kreher, D.L., & Stinson, D.R. (1999). Combinatorial algorithms: generation, 
enumeration, and search. \emph{SIGACT News}, 30(1), 33--35.

Batagelj, V., & Mrvar, A. (2001). A subquadratic triad census algorithm for 
large sparse networks with small maximum degree. \emph{Social Networks},
23(3), 237--243.

Brunson, J.C. (2015). Triadic analysis of affiliation networks. \emph{Network
Science}, 3(4), 480--508.
}
\seealso{
Other triad census functions: \code{\link{triad_census}},
  \code{\link{triad_closure_from_census}},
  \code{\link{triad_tallies}}
}
