% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule}
\alias{edgeWeight}
\alias{schedule}
\alias{shareWeight}
\title{Schedule}
\usage{
schedule(bigraph, v)

edgeWeight(graph, vp)

shareWeight(bigraph, vids)
}
\arguments{
\item{bigraph}{The ambient affiliation network}

\item{v}{A subset of actor node ids.}

\item{graph}{A one-mode network (used as the one-mode projection of bigraph)}

\item{vp}{Ids of vertex pair, the edge between which is being considered}

\item{vids}{Vertex ids, the event count among which is being considered}
}
\description{
The schedule of a subset of actor nodes in an affiliation network is the
induced subgraph on the actor nodes together with all event nodes incident
to at least two of the actor nodes.
}

