% Generated by roxygen2 (4.0.2): do not edit by hand
\name{schedule}
\alias{schedule}
\title{Schedule}
\usage{
schedule(bigraph, v)
}
\arguments{
\item{bigraph}{The ambient affiliation network}

\item{v}{A subset of actor node ids.}
}
\description{
The schedule of a subset of actor nodes in an affiliation network is the
induced subgraph on the actor nodes together with all event nodes incident
to at least two of the actor nodes.
}

