% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.r
\name{schedule}
\alias{schedule}
\title{Actors and their shared events}
\usage{
schedule(bigraph, actors = V(bigraph)[V(bigraph)$type == FALSE])
}
\arguments{
\item{bigraph}{An affiliation network.}

\item{actors}{A vector of actor nodes in \code{bigraph}.}
}
\description{
Given an affiliation network and a vector of actor node IDs, 
  produce the induced subgraph on the actor nodes together with all event 
  nodes incident to at least two of them. This is called the actors'
  *schedule*.
}
