% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/share.weight.R
\name{share.weight}
\alias{share.weight}
\title{Schedule}
\usage{
share.weight(bigraph, vids, name = "name")
}
\arguments{
\item{bigraph}{The ambient affiliation network}

\item{vids}{Vertex ids, the event count among which is being considered}

\item{v}{A subset of actor node ids.}
}
\description{
The schedule of a subset of actor nodes in an affiliation network is the
induced subgraph on the actor nodes together with all event nodes incident
to at least two of the actor nodes.
}

