% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-triad-census.R
\name{simple_triad_census}
\alias{simple.triad.census}
\alias{simple_triad_census}
\title{Simple triad census}
\usage{
simple_triad_census(graph, add.names = FALSE)

simple.triad.census(graph, add.names = FALSE)
}
\arguments{
\item{graph}{An igraph object.}

\item{add.names}{Logical; whether to label the matrix rows and columns}
}
\description{
This function inputs a simple graph and returns a vector of frequency counts 
for the four isomorphism classes of triad, which therefore sum to choose(n,
3), where n is the number of nodes. This function is needed because 
\code{triad_census}, from which in small cases a simple triad census can be 
recovered through strategic coordinate sums, returns nonsense frequencies for
many large graphs.
}

