% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tmtc2setc}
\alias{tmtc2setc}
\title{Two-mode triad census to structural equivalence triad census}
\usage{
tmtc2setc(tmtc)
}
\arguments{
\item{tmtc}{A full two-mode triad census (not necessarily labeled)}
}
\description{
Two-mode triad census to structural equivalence triad census
}

