% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transitivity.an.R
\name{transitivity.an}
\alias{transitivity.an}
\title{Affiliation network clustering coefficients}
\usage{
transitivity.an(bigraph, type = "global", stat = "clust",
  wedges.fn = injequ.wedges, vids = which(!V(bigraph)$type),
  add.names = FALSE)
}
\arguments{
\item{bigraph}{An affiliation network.}

\item{type}{Character; the type of clustering coefficient (defaults to
"global").}

\item{stat}{Character; the form of the statistic (matched to "clustering" or
"transitivity"; defaults to "clust").}

\item{wedges.fn}{The wedge function (see the entry on `wedges`)}

\item{vids}{A subset of actor node ids at which to evaluate the local
clustering coefficient}

\item{add.names}{Logical; whether to label the matrix rows and columns}
}
\description{
Each clustering coefficient can be defined as the proportion of "wedges" that
are "closed", for suitable definitions of both terms. The main function,
transitivity.an, calls one of the wedge functions and computes the
global or local clustering coefficient of the given affiliation network,
and if the local, then at the given nodes.
}
\examples{
data(ddggs.clique)
transitivity.table <- sapply(
  c(injequ.wedges, injstr.wedges, indstr.wedges),
  transitivity.an, bigraph = ddggs.clique, type = "local"
)
}

