% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad-census-an.R, R/triad-census-method.r
\name{triad_census_an}
\alias{triad_census_an}
\alias{triad.census.an}
\alias{triad_census_an}
\alias{triad_census_method}
\title{Triad census for affiliation networks}
\usage{
triad_census_an(bigraph, add.names = FALSE, verbose = FALSE)

triad.census.an(bigraph, add.names = FALSE, verbose = FALSE)

triad_census_method(bigraph, method = "batagelj_mrvar", ...,
  add.names = FALSE)
}
\arguments{
\item{bigraph}{An affiliation network.}

\item{add.names}{Logical; whether to label the rows and columns of the output
matrix.}

\item{verbose}{Logical; whether to display progress bars.}

\item{method}{Character; the triad census method to use. Currently only 
\code{"batagelj_mrvar"} is implemented. \code{"original"} calls an
inefficient but reliable original implementation in R.}

\item{...}{Additional arguments passed to the \code{method} function.}
}
\value{
A matrix counts of triad congruence classes, with row indices 
  reflecting pairwise exclusive events and column indices reflecting 
  triadwise events.
}
\description{
Given an affiliation network, tally all actor triads by 
  isomorphism class.
}
\details{
The \code{triad_census_*} functions implement the four triad
  censuses described below in matrix form.


}
\section{Triad census}{
 The function \code{\link{triad_census_an}} computes 
  the full triad census for an affiliation network. The functions 
  \code{\link{unif_triad_census}} and \code{\link{str_triad_census}} compute 
  the more compact uniformity and structural triad censuses, respectively, 
  using similar methods. The function \code{\link{simple_triad_census}} 
  computes the classical (undirected) triad census for the actor projection 
  of an affiliation network, using \code{\link[igraph]{triad_census}} but, if
  the result doesn't make sense (i.e. the sum of the entries is not the 
  number of triples of nodes), then it instead uses its own, much slower 
  method. Each of these censuses can be projected from the previous using the
  function \code{\link{project_census}}.
}

\examples{
data(women_clique)
tc <- triad_census_an(women_clique, add.names = TRUE)
tc
sum(tc) == choose(vcount(actor_projection(women_clique)), 3)
}
