% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitivity-from-census.R
\name{triad_closure_from_census}
\alias{triad_closure_from_census}
\alias{transitivity_from_census}
\alias{transitivity.census}
\alias{ftc2indequ}
\alias{ftc2allact}
\alias{ftc2injequ}
\alias{ftc2indstr}
\alias{ftc2injact}
\alias{ftc2injstr}
\title{Global triad closure from a triad census}
\usage{
triad_closure_from_census(census, measure, scheme = NULL, openFun, closedFun,
  counts = FALSE)

transitivity_from_census(census, measure, scheme = NULL, openFun, closedFun,
  counts = FALSE)

transitivity.census(census, measure, scheme = NULL, openFun, closedFun,
  counts = FALSE)

ftc2indequ(census)

ftc2allact(census)

ftc2injequ(census)

ftc2indstr(census)

ftc2injact(census)

ftc2injstr(census)
}
\arguments{
\item{census}{Numeric matrix or vector; an affiliation network triad census. 
It is treated as binary or simple if its dimensons are 4-by-2 or 4-by-1, 
respectively, unless otherwise specified by \code{scheme}; otherwise it is 
treated as full.}

\item{measure}{Character; the type of triad closure (matched to 
"watts.strogatz", "classical", "opsahl", "exclusive", "allact", "indequ", 
"indstr", "injact", "injequ", or "injstr")}

\item{scheme}{Character; the type of triad census to calculate, matched to 
\code{"full"}, \code{"binary"} (equivalently, \code{"structural"}), or 
\code{"simple"}.}

\item{openFun}{The open wedge count for a triad (ignored if \code{measure} is
not \code{NULL}).}

\item{closedFun}{The closed wedge count for a triad (ignored if
\code{measure} is not \code{NULL}).}

\item{counts}{Logical; whether to return open and closed wedge counts instead
of a ratio statistic.}
}
\description{
Given a triad census of a suitable scheme, calculate a global 
  measure of triad closure for the associated affiliation network.
}
\details{
Each global measure of triad closure can be recovered from the full 
  triad census, and some can be recovered from smaller censuses. This 
  function verifies that a given census is sufficient to recover a given 
  measure of triad closure and, if it is, returns its value.
}
