% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triadclosure-an.r
\name{triadclosure_an}
\alias{triadclosure_an}
\title{Triad closure in affiliation networks}
\usage{
triadclosure_an(bigraph, actors = V(bigraph)[V(bigraph)$type == FALSE],
  type = "global", ...)
}
\arguments{
\item{bigraph}{An affiliation network.}

\item{actors}{Actor nodes in \code{bigraph}; (re)set to \code{V(bigraph)} if 
\code{type} is \code{"global"}.}

\item{type}{The type of statistic, matched to \code{"global"}, 
\code{"local"}, or \code{"raw"}.}

\item{...}{Flavor specifications passed to \code{\link{wedges_an}}.}
}
\value{
A numeric vector of length \code{length(actors)} giving the value of 
  the specified local triad closure statistic at each actor node.

If \code{type} is \code{"global"}, the global statistic for 
  \code{bigraph}; if \code{"local"}, the local statistics for \code{actors}; 
  if \code{"raw"}, a 2-column matrix, each row of which gives the number of
  wedges and of closed wedges centered at \code{actors}.
}
\description{
Given an affiliation network and a vector of node IDs, calculate
  the (proportion of) triad closure of a specified flavor at each node. See 
  \code{\link{wedges_an}} and \code{triads_an} for details.
}
