% Generated by roxygen2 (4.0.1): do not edit by hand
\name{utc2stc}
\alias{utc2stc}
\title{Uniformity triad census to structural triad census}
\usage{
utc2stc(tc)
}
\arguments{
\item{tc}{A uniformity triad census (not necessarily labeled)}
}
\description{
Uniformity triad census to structural triad census
}

