% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wedge.dependency}
\alias{wedge.dependency}
\title{Wedge-dependency table}
\usage{
wedge.dependency(wedges)
}
\arguments{
\item{wedges}{A table of two columns: wedge count and closed wedge count}
}
\description{
Given a table of local wedge counts and closed wedge counts by node, return
a table of node counts and closed wedge proportions (mean local clustering
coefficients) by local wedge count.
}

