% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wedgeReturn.R
\name{wedgeReturn}
\alias{wedgeReturn}
\title{Compress a wedgelist into a desired statistic}
\usage{
wedgeReturn(wedges, type, add.names)
}
\arguments{
\item{wedges}{A two-column matrix of wedge and closed wedge counts centered at each of some collection of nodes.}

\item{type}{Character; the type of clustering coefficient (defaults to
"global").}
}
\value{
A transitivity statistic of type \code{type} calculated from the wedges \code{wedges}, in a format that can be fed into \code{sapply}.
}
\description{
Compress a wedgelist into a desired statistic
}

