% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{whigs}
\alias{whigs}
\title{Membership network of American Whigs.}
\format{An affiliation network; see \code{\link{is_an}}.}
\source{
Fischer (1994) *Paul Revere's Ride*
\url{https://github.com/kjhealy/revere}
}
\description{
These data record the membership of 136 colonial Americans in 5 Whig
organizations.
}
