% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{women_group}
\alias{women_group}
\title{Group of women connected by event coattendance in Old City.}
\format{An affiliation network; see \code{\link{is_an}}. Events carry a
  \code{time} attribute equal to the number of days since the previous 31
  December.}
\source{
Davis, Gardner, and Gardner (1941) *Deep South: A Social 
  Anthropological Study of Caste and Class* 
  \url{http://books.google.com/books?id=HGIdAAAAIAAJ}
}
\description{
These data record the attendance of 18 white women at 14 social events over a
9-month period.
}

